% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{ac}
\alias{ac}
\title{FUNCTION_TITLE}
\usage{
ac(
  ...,
  all = FALSE,
  commit_msg,
  path_to_local_repo = NULL,
  verbose = TRUE,
  max_mb = 50
)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{all}{PARAM_DESCRIPTION, Default: FALSE}

\item{commit_msg}{PARAM_DESCRIPTION}

\item{path_to_local_repo}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{max_mb}{PARAM_DESCRIPTION, Default: 50}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
