% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{acp}
\alias{acp}
\title{Add, Commit, and Push}
\usage{
acp(
  commit_msg,
  ...,
  path = getwd(),
  pattern = NULL,
  all.files = FALSE,
  recursive = FALSE,
  ignore.case = FALSE,
  include.dirs = FALSE,
  no.. = FALSE,
  max_mb = 50,
  remote_name = "origin",
  remote_branch = "master",
  verbose = TRUE
)
}
\description{
Add, Commit, and Push
}
