% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{FUNCTION_TITLE}
\usage{
add(..., all = TRUE, path_to_local_repo = NULL, max_mb = 50)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{all}{PARAM_DESCRIPTION, Default: FALSE}

\item{path_to_local_repo}{PARAM_DESCRIPTION, Default: NULL}

\item{max_mb}{PARAM_DESCRIPTION, Default: 50}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
 \code{\link[purrr]{map}}
}
