% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_all}
\alias{add_all}
\title{Add All Files to a Commit}
\usage{
add_all(path_to_local_repo = NULL)
}
\arguments{
\item{path_to_local_repo}{Full path to local repository. If NULL, adds all the files}
}
\description{
Add All Files to a Commit
}
