% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_all0.R
\name{add_all0}
\alias{add_all0}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
add_all0(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
