% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_all_in_outside_repo.R
\name{add_all_in_outside_repo}
\alias{add_all_in_outside_repo}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
add_all_in_outside_repo(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
