% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_all_this_repo.R
\name{add_all_this_repo}
\alias{add_all_this_repo}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
add_all_this_repo()
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
