% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commit_all.R
\name{add_commit_all}
\alias{add_commit_all}
\title{Add and commit all changes in a local repo}
\usage{
add_commit_all(
  path_to_local_repo = NULL,
  commit_message = NULL,
  description = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path_to_local_repo}{full path to local repository where the add and commit all will be performed}

\item{commit_message}{If NULL, automatically creates a message in the format of "add/modify {filename} as written in {R script path}"}

\item{description}{additional optional description}
}
\description{
Add and commit all changes in a local repo
}
