% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commit_all_deltas_in_path.R
\name{add_commit_all_deltas_in_path}
\alias{add_commit_all_deltas_in_path}
\title{Adds and commits all changed files in a given path}
\usage{
add_commit_all_deltas_in_path(path, commit_message = NULL, description = NULL)
}
\description{
Adds and commits all changed files in a given path
}
