% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commit_all_deltas_in_wd.R
\name{add_commit_all_deltas_in_wd}
\alias{add_commit_all_deltas_in_wd}
\title{Adds and commits all changed files in the current wd}
\usage{
add_commit_all_deltas_in_wd(commit_message = NULL, description = NULL)
}
\description{
Adds and commits all changed files in the current wd
}
