% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commit_all_modified_files.R
\name{add_commit_all_modified_files}
\alias{add_commit_all_modified_files}
\title{Add and Commit Modified Files}
\usage{
add_commit_all_modified_files(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{path to local repo}
}
\description{
Adds and commits all modified files one-by-one in the local repo with
a commit message that states that the file was committed automatically using this function.
}
