% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commit_all_this_repo.R
\name{add_commit_all_this_repo}
\alias{add_commit_all_this_repo}
\title{Add and commit all chnages in repo belonging to the current working directory}
\usage{
add_commit_all_this_repo(commit_message, description = NULL)
}
\arguments{
\item{commit_message}{commit message}

\item{description}{description, if desired. Defaults to NULL.}
}
\description{
Add and commit all chnages in repo belonging to the current working directory
}
