% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commit_path_deltas.R
\name{add_commit_path_deltas}
\alias{add_commit_path_deltas}
\title{Adds and commits all changed files in a given path}
\usage{
add_commit_path_deltas(path, commit_message = NULL, description = NULL)
}
\description{
This function combines outputs for
}
