% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_commit_some}
\alias{add_commit_some}
\title{Add and Commit Some Files}
\usage{
add_commit_some(
  filenames,
  commit_message = NULL,
  description = NULL,
  verbose = TRUE,
  path_to_local_repo = NULL
)
}
\arguments{
\item{filenames}{names of files in the local repository path to be committed}

\item{commit_message}{defaults to NULL, where the commit_message is a "change to {filename} in {path to present R script}"}

\item{path_to_local_repo}{full path to local repository to be pushed}
}
\description{
Add and Commit Some Files
}
