% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_descriptor.R
\name{add_descriptor}
\alias{add_descriptor}
\title{Builds the "select" part of a query.}
\usage{
add_descriptor(query, name, formula)
}
\arguments{
\item{query}{a list with elements of the query}

\item{name}{name of variable}

\item{description}{formula}
}
\description{
Builds the "select" part of a query.
}
\examples{
add_triplets(subject="wd:Q331676",verb="wdt:P1843",object="?statement") \%>\%
add_descriptor(name="?lang", formula="LANG(?statement)") \%>\%
send()
}
