% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_file.R
\name{add_file}
\alias{add_file}
\title{Add a Single File to a Commit}
\usage{
add_file(path_to_local_repo = NULL, file, verbose = TRUE)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}
}
\description{
This function is a primitive function to wrap add_some() around.
}
