% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_some_this_repo.R
\name{add_some_this_repo}
\alias{add_some_this_repo}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
add_some_this_repo(filenames)
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
