% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated1.R
\name{addToGitIgnore}
\alias{addToGitIgnore}
\title{Add an Entry to .gitignore}
\usage{
addToGitIgnore(..., commit = TRUE, path_to_local_repo = NULL)
}
\description{
This function adds an entry to the .gitignore file in the repo. If one is not present, it is created.
}
