% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_file_deltas_in_wd.R
\name{all_file_deltas_in_wd}
\alias{all_file_deltas_in_wd}
\title{Get a vector of changed files in the current working directory}
\usage{
all_file_deltas_in_wd()
}
\description{
Get a vector of changed files in the current working directory
}
