% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{as_values}
\alias{as_values}
\title{transforms subject-verb-object arguments into a vector of values if needed}
\usage{
as_values(string)
}
\description{
transforms subject-verb-object arguments into a vector of values if needed
}
\examples{
object="wd:Q456|wd:Q23482"
}
