% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{branch}
\alias{branch}
\title{Git Branch}
\usage{
branch(verbose = TRUE, path = getwd())
}
\value{
Invisibly returns a vector of branch names
}
\description{
Print all the branches and the current branch in a GitHub repository
}
