% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-browse_github_functions.R
\name{browse_gh_pages}
\alias{browse_gh_pages}
\title{Open a GitHub Page in a Browser}
\usage{
browse_gh_pages(github_username, repo)
}
\arguments{
\item{github_username}{GitHub username}

\item{repo}{Repository belonging to the GitHub user}
}
\value{
If a GitHub Page exists, it is opened in the browser. Otherwise, a message is returned in the console that states that a GitHub Page does not exist.
}
\description{
These functions open up GitHub urls related to a repo in the browser returned by \code{\link{get_remote_urls}}.
}
\details{
This function opens the GitHub Page URL returned by \code{\link{get_remote_urls}} if a GitHub Page exists for the repo.
}
\seealso{
\code{\link[dplyr]{filter}}
 \code{\link[secretary]{typewrite}}
 \code{\link[utils]{browseURL}}

Other browse github: 
\code{\link{browse_repo_issues}()}
}
\concept{browse github}
