% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{browse_gh_pages2}
\alias{browse_gh_pages2}
\title{Open a GitHub Page in a Browser}
\usage{
browse_gh_pages2(github_username, repo)
}
\value{
If a GitHub Page exists, it is opened in the browser. Otherwise, a message is returned in the console that states that a GitHub Page does not exist.
}
\description{
Open a GitHub Page in a Browser
}
\details{
This function opens the GitHub Page URL returned by \code{\link{get_remote_urls}} if a GitHub Page exists for the repo.
}
\seealso{
\code{\link[dplyr]{filter}}
 \code{\link[secretary]{typewrite}}
 \code{\link[utils]{browseURL}}
}
\concept{browse github}
