% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_gh.R
\name{browse_github}
\alias{browse_github}
\title{Browse GitHub}
\arguments{
\item{github_user}{GitHub username}

\item{repo}{Repository belonging to the GitHub user}

\item{issue_no}{Issue Number}
}
\description{
Browse GitHub
}
