% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-browse_github_functions.R
\name{browse_repo_issues}
\alias{browse_repo_issues}
\alias{browse_repo}
\title{Open a GitHub Repo URL in a Browser}
\usage{
browse_repo_issues(github_username, repo)

browse_repo(github_username, repo)
}
\arguments{
\item{github_username}{GitHub username}

\item{repo}{Repository belonging to the GitHub user}
}
\description{
These functions open up GitHub urls related to a repo in the browser returned by \code{\link{get_remote_urls}}.
}
\seealso{
\code{\link[utils]{browseURL}}

\code{\link[utils]{browseURL}}

Other browse github: 
\code{\link{browse_gh_pages}()}
}
\concept{browse github}
