% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_classes.R
\name{build_graph_classes}
\alias{build_graph_classes}
\title{Build the graph of sub/superclasses}
\usage{
build_graph_classes(id)
}
\arguments{
\item{id}{the id of class}
}
\description{
Build the graph of sub/superclasses
}
\examples{
\dontrun{
tib_g=build_graph_classes("wd:Q627272")
}
}
