% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_part_body.R
\name{build_part_body}
\alias{build_part_body}
\title{Builds the "body" part of a query.}
\usage{
build_part_body(
  query = NA,
  subject,
  verb,
  object,
  required = TRUE,
  within_box = c(NA, NA),
  within_distance = c(NA, NA)
)
}
\arguments{
\item{query}{a list with elements of the query}

\item{subject}{an anonymous variable (for instance, and by default, "?subject") or item (for instance "wd:Q456"))}

\item{verb}{the property (for instance "wdt:P190")}

\item{object}{an anonymous variable (for instance, and by default, "?object") or item (for instance "wd:Q456"))}

\item{required}{whether the existence of a value for the triple is required or not (defaults to TRUE).
If set to FALSE, then other triples in the query are returned even if this particular triple is missing)}

\item{within_box}{if provided, north-west and south-east coordinates of bounding box for the triple query.}

\item{within_distance}{if provided, north-west and south-east coordinates of bounding box for the triple query.}
}
\description{
Builds the "body" part of a query.
}
