% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parts.R
\name{build_part_filter}
\alias{build_part_filter}
\title{Builds the "filter" part of a query.}
\usage{
build_part_filter(query = NA, expressions)
}
\arguments{
\item{query}{a list with elements of the query*}

\item{expressions}{the vector of filtering expressions to apply}
}
\description{
Builds the "filter" part of a query.
}
