% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_part_limit.R
\name{build_part_limit}
\alias{build_part_limit}
\title{Builds the "limit" part of a query.}
\usage{
build_part_limit(query, limit = NA)
}
\arguments{
\item{query}{a list with elements of the query}
}
\description{
Builds the "limit" part of a query.
}
