% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_part_prefixes.R
\name{build_part_prefixes}
\alias{build_part_prefixes}
\title{Builds the "prefixes" part of a query.}
\usage{
build_part_prefixes(query = NA, prefixes = NULL)
}
\arguments{
\item{query}{a list with elements of the query}
}
\description{
Builds the "prefixes" part of a query.
}
