% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parts.R
\name{build_part_service}
\alias{build_part_service}
\title{Builds the "service" part of a query.}
\usage{
build_part_service(query = NA, language = "en")
}
\arguments{
\item{query}{a list with elements of the query}
}
\description{
Builds the "service" part of a query.
}
\examples{
build_part_service(query=NA, language="fr")
}
