% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_push_site.R
\name{build_push_site}
\alias{build_push_site}
\title{Build and Push docs/ dir For GitHub Pages}
\usage{
build_push_site(
  commit_message = "update GitHub Page",
  lazy = TRUE,
  preview = FALSE,
  devel = TRUE,
  ...
)
}
\description{
This function automatically documents, pushes, and installs a package.
}
\seealso{
\code{\link[usethis]{use_pkgdown}}
 \code{\link[pkgdown]{build_site}}
}
\keyword{internal}
