% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sparql.R
\name{build_sparql}
\alias{build_sparql}
\title{Assemble query parts into a proper SPARQL query}
\usage{
build_sparql(query, endpoint = "Wikidata")
}
\arguments{
\item{query}{a list with elements of the query}

\item{endpoint}{SPARQL endpoint to send the query to}
}
\description{
Assemble query parts into a proper SPARQL query
}
\examples{
add_triplets(subject="?city",verb="wdt:P31",object="wd:Q515") \%>\%
add_triplets(subject="?city",verb="wdt:P1082",object="?pop", label="?city", language="en") \%>\%
spq_head() \%>\%
build_sparql() \%>\%
cat()
}
