% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeRemoteURL.R
\name{changeRemoteURL}
\alias{changeRemoteURL}
\title{FUNCTION_TITLE}
\usage{
changeRemoteURL(path_to_local_repo, pattern, replacement, remote = "origin")
}
\arguments{
\item{path_to_local_repo}{PARAM_DESCRIPTION}

\item{pattern}{PARAM_DESCRIPTION}

\item{replacement}{PARAM_DESCRIPTION}

\item{remote}{PARAM_DESCRIPTION, Default: 'origin'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glitter]{remote_url}}
 \code{\link[stringr]{str_replace}}
}
