% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repo_status.R
\name{check_repo_status}
\alias{check_repo_status}
\title{Check whether a commit is needed or not for a given repo}
\usage{
check_repo_status(path_to_local_repo)
}
\description{
Check whether a commit is needed or not for a given repo
}
