% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{checkout}
\alias{checkout}
\title{Checkout a Branch}
\usage{
checkout(path_to_local_repo = NULL, branch)
}
\arguments{
\item{branch}{Name of branch to checkout}
}
\description{
Print all the branches and the current branch in a GitHub repository
}
