% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout_functions.R
\name{checkout_branch}
\alias{checkout_branch}
\title{Checkout a Branch}
\usage{
checkout_branch(branch, path_to_local_repo = NULL)
}
\arguments{
\item{branch}{Name of the new branch}
}
\description{
Checkout a Branch
}
