% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{checkout_new_branch}
\alias{checkout_new_branch}
\title{Checkout a New Branch}
\usage{
checkout_new_branch(new_branch, path_to_local_repo = NULL)
}
\arguments{
\item{new_branch}{Name of the new branch}
}
\description{
Print all the branches and the current branch in a GitHub repository
}
