% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wikidata_table.R
\name{clean_wikidata_table}
\alias{clean_wikidata_table}
\title{Clean Wikidata results by removing the "http://www.wikidata.org/entity/" prefix in all columns}
\usage{
clean_wikidata_table(wikitib)
}
\arguments{
\item{wikitib}{a Wikidata results tibble}
}
\description{
Clean Wikidata results by removing the "http://www.wikidata.org/entity/" prefix in all columns
}
\examples{
metro_query='SELECT ?item ?itemLabel ?coords
{
 ?item wdt:P361 wd:Q1552;
 wdt:P625 ?coords .
 SERVICE wikibase:label { bd:serviceParam wikibase:language "en" . }
} ORDER BY ?itemLabel
'
tib=send_sparql(query=metro_query)
clean_tib=clean_wikidata_table(tib)
clean_tib \%>\% head()
}
