% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{closeIssue}
\alias{closeIssue}
\title{(Deprecated) Close a GitHub Issue}
\usage{
closeIssue(path = getwd(), issueNo, closureMsg)
}
\description{
(Deprecated) This function closes a GitHub Issue from the R console.
}
