% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.R
\name{com}
\alias{com}
\title{FUNCTION_TITLE}
\usage{
com(commit_msg, path_to_local_repo = NULL, verbose = TRUE)
}
\arguments{
\item{commit_msg}{PARAM_DESCRIPTION}

\item{path_to_local_repo}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[secretary]{typewrite_bold}},\code{\link[secretary]{character(0)}},\code{\link[secretary]{typewrite_italic}}
}
