% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit}
\alias{commit}
\title{Push a Local Repository}
\usage{
commit(
  path_to_local_repo = NULL,
  commit_message,
  description = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}

\item{commit_message}{message to be included in the commit}
}
\description{
Push a Local Repository
}
