% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit0.R
\name{commit0}
\alias{commit0}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
commit0(path_to_local_repo, commit_message = NULL, description = NULL)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}

\item{commit_message}{message to be included in the commit}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
