% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_log.R
\name{commit_log}
\alias{commit_log}
\title{See Commit Log}
\usage{
commit_log(path = getwd(), verbose = TRUE, remove = TRUE)
}
\description{
See the history of all the commits that have been made in the console. The history is also parsed and invisibly returned as a dataframe with `sha` and `commit_msg` fields.
}
