% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_outside_repo.R
\name{commit_outside_repo}
\alias{commit_outside_repo}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
commit_outside_repo(path_to_local_repo, commit_message, description = NULL)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}

\item{commit_message}{message to be included in the commit}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
