% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_this_repo.R
\name{commit_this_repo}
\alias{commit_this_repo}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
commit_this_repo(commit_message, description = NULL)
}
\arguments{
\item{commit_message}{message to be included in the commit}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
