% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated1.R
\name{commitHistoryGraph}
\alias{commitHistoryGraph}
\title{See the history of your commits in a graph}
\usage{
commitHistoryGraph(path_to_local_repo = NULL, verbose = TRUE)
}
\description{
See the history of your commits in a graph
}
