% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_classes.R
\name{count_items}
\alias{count_items}
\title{Get counts of items directly in classes}
\usage{
count_items(classes)
}
\arguments{
\item{classes}{the vector wikidata classes}
}
\description{
Get counts of items directly in classes
}
\examples{
count_items("wd:Q7930989")
}
