% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_triple.R
\name{decompose_triple}
\alias{decompose_triple}
\title{Get elements subject, verb, object of a triple as a list.}
\usage{
decompose_triple(triple, subject, verb, object)
}
\arguments{
\item{triple}{the triple statement (replaces arguments subject verb and object)}

\item{subject}{an anonymous variable (for instance, and by default, "?subject") or item (for instance "wd:Q456"))}

\item{verb}{the property (for instance "wdt:P190")}

\item{object}{an anonymous variable (for instance, and by default, "?object") or item (for instance "wd:Q456"))}
}
\description{
Get elements subject, verb, object of a triple as a list.
}
