% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose-triple-pattern.R
\name{decompose_triple_pattern}
\alias{decompose_triple_pattern}
\title{Get elements subject, verb, object of a triple pattern as a list.}
\usage{
decompose_triple_pattern(triple_pattern, subject, verb, object)
}
\arguments{
\item{triple_pattern}{the triple pattern statement (replaces arguments subject verb and object)}

\item{subject}{an anonymous variable (for instance, and by default, "?subject") or item (for instance "wd:Q456"))}

\item{verb}{the property (for instance "wdt:P190")}

\item{object}{an anonymous variable (for instance, and by default, "?object") or item (for instance "wd:Q456"))}
}
\description{
Get elements subject, verb, object of a triple pattern as a list.
}
