% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_triplet.R
\name{decompose_triplet}
\alias{decompose_triplet}
\title{Get elements subject, verb, object of a triplet as a list.}
\usage{
decompose_triplet(triplet, subject, verb, object)
}
\arguments{
\item{triplet}{the triplet statement (replaces arguments subject verb and object)}

\item{subject}{an anonymous variable (for instance, and by default, "?subject") or item (for instance "wd:Q456"))}

\item{verb}{the property (for instance "wdt:P190")}

\item{object}{an anonymous variable (for instance, and by default, "?object") or item (for instance "wd:Q456"))}
}
\description{
Get elements subject, verb, object of a triplet as a list.
}
