% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupeGitIgnore.R
\name{dedupeGitIgnore}
\alias{dedupeGitIgnore}
\title{Deduplicate .gitignore}
\usage{
dedupeGitIgnore(path_to_local_repo = NULL, commit = TRUE)
}
\description{
This function reads the .gitignore file, removes duplicates, and overwrites the gitignore file.
}
