% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_branch.R
\name{delete_branch}
\alias{delete_branch}
\title{Delete a Branch}
\usage{
delete_branch(path_to_local_repo = NULL, branch)
}
\arguments{
\item{branch}{Name of branch to checkout}
}
\description{
Delete a branch
}
