% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{delete_local_tag}
\alias{delete_local_tag}
\title{FUNCTION_TITLE}
\usage{
delete_local_tag(path = getwd(), tag)
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: getwd()}

\item{tag}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[git2r]{tag_delete}}
}
