% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{delete_remote_branch}
\alias{delete_remote_branch}
\title{Delete a Remote Branch}
\usage{
delete_remote_branch(path = getwd(), remote_branch)
}
\arguments{
\item{remote_branch}{Remote branch.}
}
\description{
Delete a remote branch of a given local clone of a repo.
}
