% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{delete_remote_tag}
\alias{delete_remote_tag}
\title{FUNCTION_TITLE}
\usage{
delete_remote_tag(path = getwd(), tag, remote_name = "origin")
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: getwd()}

\item{tag}{PARAM_DESCRIPTION}

\item{remote_name}{PARAM_DESCRIPTION, Default: 'origin'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
