% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleted_files.R
\name{deleted_files}
\alias{deleted_files}
\title{Get a list of deleted files in a repo}
\usage{
deleted_files(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{path to local repo}
}
\description{
Get a list of deleted files in a repo
}
