% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleted_files0.R
\name{deleted_files0}
\alias{deleted_files0}
\title{Get a list of deleted files in a repo}
\usage{
deleted_files0(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{path to local repo}
}
\value{
The Git Status message and deleted files are printed in the console while the vector of the deleted filenames are invisibly returned.
}
\description{
Lists all the deleted files according to the git status message returned by the status function
}
