% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_gh_pages}
\alias{deploy_gh_pages}
\title{Re-Build and Push GitHub Pages}
\usage{
deploy_gh_pages(
  path = getwd(),
  examples = TRUE,
  run_dont_run = FALSE,
  seed = 1014,
  lazy = TRUE,
  override = list(),
  preview = FALSE,
  devel = TRUE,
  new_process = !devel,
  install = !devel
)
}
\description{
This unlinks the docs subdirectory, runs devtools document function, and writes a new docs/ directory using the pkgdown build_site function.
}
\seealso{
\code{\link[usethis]{use_pkgdown}}
 \code{\link[pkgdown]{build_site}}
}
