% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_pkg}
\alias{deploy_pkg}
\title{Deploy a Package}
\usage{
deploy_pkg(
  commit_msg = "deploy pkg using glitter",
  tag = NULL,
  remote_name = "origin",
  remote_branch = "master",
  install = TRUE,
  reset = FALSE,
  build_vignettes = FALSE,
  path = getwd(),
  ref = "HEAD",
  git = c("auto", "git2r", "external"),
  dependencies = NA,
  upgrade = "never",
  force = FALSE,
  quiet = FALSE,
  build = TRUE,
  build_opts = c("--no-resave-data", "--no-manual", "--no-build-vignettes"),
  build_manual = FALSE,
  repos = getOption("repos"),
  type = getOption("pkgType")
)
}
\arguments{
\item{commit_msg}{commit message}

\item{install}{If TRUE, installs the package after the changes are pushed to the remote, Default: TRUE}

\item{reset}{If TRUE, restart R after installation is complete. Default: TRUE.}

\item{description}{description to extend the commit message if desired, Default: NULL}
}
\value{
A freshly packed local package committed to the remote that is by default also installed with vignettes, if applicable.
}
\description{
Deploy a Package
}
\seealso{
\code{\link[devtools]{document}},\code{\link[devtools]{build_vignettes}},\code{\link[devtools]{remote-reexports}}
}
