% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroy.R
\name{destroy}
\alias{destroy}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
destroy(path_to_local_repo, permanently_remove_this_file)
}
\arguments{
\item{permanently_remove}{filename of the file to permanently remove all traces of}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
