% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_push_install_package.R
\name{doc_push_install_package}
\alias{doc_push_install_package}
\title{Document, push changes, and install a public GitHub Package
This function automatically documents, pushes, and installs a package, assuming that the basename fo the working directory is the same as the repo as in patelm9/{repo}. If the URL of the GitHub remote belongs to MSKCC, the package is instead installed using a Git hyperlink.}
\usage{
doc_push_install_package(
  commit_message,
  description = NULL,
  github_pages = TRUE
)
}
\description{
Document, push changes, and install a public GitHub Package
This function automatically documents, pushes, and installs a package, assuming that the basename fo the working directory is the same as the repo as in patelm9/{repo}. If the URL of the GitHub remote belongs to MSKCC, the package is instead installed using a Git hyperlink.
}
