% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docPushBuildInstall.R
\name{docPushBuildInstall}
\alias{docPushBuildInstall}
\title{Document, Push, Install, and Build a Site}
\usage{
docPushBuildInstall(
  commit_message,
  description = NULL,
  install = TRUE,
  build = TRUE,
  verbose = TRUE,
  ...
)
}
\description{
Document, Push, Install, and Build a Site
}
\seealso{
\code{\link[devtools]{document}}, \code{\link[devtools]{remote-reexports}}
}
\keyword{internal}
