% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{docPushInstall}
\alias{docPushInstall}
\title{Document, Push, and Install a GitHub Package}
\usage{
docPushInstall(
  commit_message = "update documentation",
  description = NULL,
  install = TRUE,
  reset = FALSE,
  has_vignettes = FALSE,
  path = getwd()
)
}
\arguments{
\item{commit_message}{commit message}

\item{description}{description to extend the commit message if desired, Default: NULL}

\item{install}{If TRUE, installs the package after the changes are pushed to the remote, Default: TRUE}

\item{reset}{If TRUE, restart R after installation is complete. Default: TRUE.}

\item{has_vignettes}{If TRUE, vignettes in the vignette/ subdir are built, pushed, and also built upon installation. Default: TRUE}
}
\value{
A freshly packed local package committed to the remote that is by default also installed with vignettes, if applicable.
}
\description{
This function automatically documents, pushes, and installs a package, assuming that the basename fo the working directory is the same as the repo as in patelm9/{repo}. If the URL of the GitHub remote belongs to MSKCC, the package is instead installed using a Git hyperlink.
}
\seealso{
\code{\link[devtools]{document}},\code{\link[devtools]{build_vignettes}},\code{\link[devtools]{remote-reexports}}
}
\keyword{internal}
