% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_to_commit.R
\name{files_to_commit}
\alias{files_to_commit}
\title{Get a list of both untracked and modified files a local git repo}
\usage{
files_to_commit(path_to_local_repo)
}
\description{
This function gets all the modified, untracked, and deleted files in the local repo at path_to_local_repo
}
\keyword{internal}
