% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsFilesToCommit.R
\name{filesToCommit}
\alias{filesToCommit}
\title{List Files To Commit}
\usage{
filesToCommit(path_to_local_repo = NULL, label = FALSE)
}
\description{
This function combines all the Unstaged and Untracked files in a repo.
}
