% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_repo.R
\name{filter_repo}
\alias{filter_repo}
\title{Destroy all traces of a file accidentally pushed to a public repo. This requires creating a fresh local clone of the repo to rewrite history. This function does not alter the repo from the command line, but rather returns the CLI commands to do so}
\usage{
filter_repo(path_to_local_repo_clone, destroy_this_file)
}
\arguments{
\item{path_to_local_repo_clone}{path to a clone of the local repo}
}
\description{
Destroy all traces of a file accidentally pushed to a public repo. This requires creating a fresh local clone of the repo to rewrite history. This function does not alter the repo from the command line, but rather returns the CLI commands to do so
}
\keyword{internal}
