% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_for_cli.R
\name{format_for_cli}
\alias{format_for_cli}
\title{Subsitute spaces and special characters with the backslash for a CLI command}
\usage{
format_for_cli(vector)
}
\description{
Subsitute spaces and special characters with the backslash for a CLI command
}
