% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatCli}
\alias{formatCli}
\title{Subsitute spaces and special characters with the backslash for a CLI command}
\usage{
formatCli(vector)
}
\description{
Subsitute spaces and special characters with the backslash for a CLI command
}
