% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_claim}
\alias{get_claim}
\title{Get claims regarding one Wikidata thing}
\usage{
get_claim(id, property_name = "wd:P31")
}
\arguments{
\item{id}{a Wikidata ID, either of an item ("Qxxxxx") or of a property ("Pxxxxx"), or the item itself.}

\item{property_name}{the name of property to get}
}
\description{
Get claims regarding one Wikidata thing
}
\examples{
get_claim("wd:Q188743", "wd:P625")
}
