% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_claims}
\alias{get_claims}
\title{Get claims regarding one Wikidata thing}
\usage{
get_claims(id, with_labels = FALSE)
}
\arguments{
\item{id}{a Wikidata ID, either of an item ("wd:Qxxxxx") or of a property ("wdt:Pxxxxx"), or the item itself.}
}
\description{
Get claims regarding one Wikidata thing
}
\examples{
get_claims("wd:Q431603")
}
