% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_description}
\alias{get_description}
\title{Get description of Wikidata thing}
\usage{
get_description(id, language = "en")
}
\arguments{
\item{id}{a Wikidata ID, either of an item ("Qxxxxx") or of a property ("Pxxxxx"), or the item itself.}

\item{language}{language of description, defaults to English ("en")}
}
\description{
Get description of Wikidata thing
}
\examples{
get_description("wd:Q431603")
get_description("wd:Q431603", language="es")
}
