% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description_links.R
\name{get_description_links}
\alias{get_description_links}
\title{FUNCTION_TITLE}
\usage{
get_description_links(github_user, repo)
}
\arguments{
\item{github_user}{PARAM_DESCRIPTION}

\item{repo}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
