% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_label}
\alias{get_label}
\title{Get label of Wikidata thing}
\usage{
get_label(id, language = "en")
}
\arguments{
\item{id}{a Wikidata ID, either of an item ("Qxxxxx") or of a property ("Pxxxxx"), or the item itself.}

\item{language}{language of label, defaults to English ("en")}
}
\description{
Get label of Wikidata thing
}
\examples{
get_label("wd:Q431603")
get_label("wd:Q431603", language="fr")
}
