% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_one_claim}
\alias{get_one_claim}
\title{Format information about one claim (for use in get_claims)}
\usage{
get_one_claim(res)
}
\arguments{
\item{res}{result}
}
\description{
Format information about one claim (for use in get_claims)
}
