% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api_functions.R
\name{get_remote_repos}
\alias{get_remote_repos}
\title{Make an API Call for a GitHub User's Remote Repositories}
\usage{
get_remote_repos(github_username)
}
\arguments{
\item{github_username}{GitHub username. Enterprise GitHub is not supported.}
}
\value{
A list of length 2. List `SETTINGS` is a list of all the values in the parsed dataframe that remained constant in all rows such as all username-related (aka `owner`) fields. All other fields are maintained in the dataframe format and returned in `REPOS`.
}
\description{
This function makes an API call to the GitHub to retrieve all the repository information for a given GitHub User. All native fields in the JSON response are preserved with the additions "pages_url" and "issues_page_url" fields. The "pages_url" field is where a URL to the GitHub Pages for the repo is derived from the `name` field if the `has_pages` field is TRUE after the response is parsed into a dataframe. The same is done with the "issues_page_url" path except that no logic is applied since all GitHub repos will have an Issues page by default.
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
 \code{\link[purrr]{map}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{distinct}}
 \code{\link[rubix]{normalize_all_to_na}},\code{\link[rubix]{deselect_if_all_same}}
}
\concept{github api}
