% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{get_repos}
\alias{get_repos}
\title{Make an API Call for a GitHub User's Remote Repositories}
\usage{
get_repos(github_user, per_page = 100, page = 1)
}
\arguments{
\item{github_user}{GitHub username. Enterprise GitHub is not supported.}
}
\value{
JSON response parsed into a tibble.
}
\description{
Make an API call to the GitHub API to retrieve all the repository information for a itHub User. All native fields in the JSON response are preserved with all the url fields returned with a terminal curly bracketed field such as "{/id}" are mutated with a corresponding "page" field with that terminal pattern removed to directly open the path in the browser if desired. Additionally, an "issues_page_url" and a "pages_url" fields are added for a direct url to the repository's issues page and GitHub Pages site, respectively. The url to the GitHub Pages is derived on that condition that the `has_pages` field is TRUE for that repository. The value in this field is `NA` if a GitHub Page does not exist for the repo. The same is done with the "issues_page_url" path except that no logic is applied since all GitHub repos will have an Issues page by default.
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{http_error}},\code{\link[httr]{status_code}},\code{\link[httr]{content}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[tibble]{rownames}},\code{\link[tibble]{as_tibble}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{mutate}}
 \code{\link[stringr]{str_replace}}
}
\concept{github api functions}
