% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{get_tags}
\alias{get_tags}
\title{FUNCTION_TITLE}
\usage{
get_tags(path = getwd())
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: getwd()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[git2r]{tags}}
}
