% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_thing}
\alias{get_thing}
\title{Get Wikidata item or property
Get unformatted info from Wikidata based on an item or property id.}
\usage{
get_thing(id)
}
\arguments{
\item{id}{a Wikidata ID, either of an item ("wd:Qxxxxx") or of a property ("wd:Pxxxxx")}
}
\description{
Get Wikidata item or property
Get unformatted info from Wikidata based on an item or property id.
}
\examples{
get_thing("wd:Q431603")
}
