% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{get_varformula}
\alias{get_varformula}
\title{get full specification of variables (as in SELECT part of SPARQL query) based on variable names}
\usage{
get_varformula(selected)
}
\arguments{
\item{vars}{the selected variables (formula and name)}
}
\description{
get full specification of variables (as in SELECT part of SPARQL query) based on variable names
}
\examples{
glitter:::get_varformula(c("?author","?document","(year(?date) AS ?year)"))
}
