% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_this_repo.R
\name{git_pull_this_repo}
\alias{git_pull_this_repo}
\title{Pull a GitHub repo based on the path to local repo}
\usage{
git_pull_this_repo()
}
\description{
Pull a GitHub repo based on the path to local repo
}
