% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_url.R
\name{git_remote_url}
\alias{git_remote_url}
\title{Get Git Remote URL}
\usage{
git_remote_url(remote_name = "origin")
}
\arguments{
\item{remote_name}{Name of the remote. Defaults to "Origin".}
}
\value{
url of the remote as a string
}
\description{
Get Git Remote URL
}
