% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glitter-package.R
\docType{package}
\name{glitter-package}
\alias{glitter}
\alias{glitter-package}
\title{glitter: glitter makes SPARQL}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package aims at writing and sending SPARQL queries. It makes the exploration and use of Linked Open Data (Wikidata in particular) easier for those who do not know SPARQL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lvaudor.github.io/glitter/}
  \item \url{https://github.com/lvaudor/glitter}
  \item Report bugs at \url{https://github.com/lvaudor/glitter/issues}
}

}
\author{
\strong{Maintainer}: Lise Vaudor \email{lise.vaudor@ens-lyon.fr} (\href{https://orcid.org/0000-0003-1844-6425}{ORCID})

Authors:
\itemize{
  \item Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
}

}
\keyword{internal}
