% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{hasRemote}
\alias{hasRemote}
\title{Does the GitHub Repo Have a Remote?}
\usage{
hasRemote(path = getwd(), remote = "origin")
}
\arguments{
\item{path}{full path to local repository to be pushed}
}
\value{
TRUE if the path has a remote url.
}
\description{
Does the GitHub Repo Have a Remote?
}
