% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore_lfs_in_this_repo.R
\name{ignore_lfs_in_this_repo}
\alias{ignore_lfs_in_this_repo}
\title{Add files greater than indicated size in working directory to .gitignore}
\usage{
ignore_lfs_in_this_repo(max_size_in_gb = 1.8)
}
\arguments{
\item{max_size_in_gb}{files sized greater than this in gigabytes will be ignored in git}
}
\value{
updated .gitignore in local repository that is pushed to master origin
}
\description{
Add files greater than indicated size in working directory to .gitignore
}
