% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs.R
\name{install_lfs}
\alias{install_lfs}
\title{Install LFS}
\usage{
install_lfs(path = getwd(), lfs_ext = c("xlsx", "csv", "RData"))
}
\arguments{
\item{lfs_ext}{extensions of files that will be large and will need to be tracked using the lfs}
}
\description{
Install LFS
}
