% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{interpret_svo}
\alias{interpret_svo}
\title{interprets if element is an R code inclusion of the type {...}#' @param string a string}
\usage{
interpret_svo(string)
}
\description{
interprets if element is an R code inclusion of the type {...}#' @param string a string
}
\examples{
truc="0000000121012885"
glitter:::interpret_svo("{truc}") #"0000000121012885"
}
