% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_git_repo.R
\name{is_git_repo}
\alias{is_git_repo}
\title{Is the path a git repo?}
\usage{
is_git_repo(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}
}
\value{
If the git message is of a length greater than 0, it is returned as a character vector and also printed in the console
}
\description{
Is the path a git repo?
}
\keyword{internal}
