% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_prefix_known}
\alias{is_prefix_known}
\title{Checks whether the element has a known prefix}
\usage{
is_prefix_known(prefixes_used, prefixes_known, endpoint = "Wikidata")
}
\arguments{
\item{prefixes_used}{a string or vector of strings}

\item{prefixes_known}{a tibble detailing known prefixes}

\item{endpoint}{the endpoint considered (defaults to Wikidata)}
}
\description{
Checks whether the element has a known prefix
}
\examples{
glitter:::is_prefix_known(prefixes_used=c("wd","wdt"),prefixes_known=usual_prefixes,endpoint="Wikidata") # TRUE
glitter:::is_prefix_known("blop:blabla", prefixes_known=usual_prefixes,endpoint="other") #returns error message
}
