% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_svo_correct}
\alias{is_svo_correct}
\title{Checks whether subject/verb/object is stated correctly}
\usage{
is_svo_correct(vstring)
}
\arguments{
\item{vstring}{a string or vector of strings}
}
\description{
Checks whether subject/verb/object is stated correctly
}
\examples{
is_svo_correct("elem") #FALSE
is_svo_correct("a") #TRUE
is_svo_correct("is") #TRUE
is_svo_correct("'0000000121012885'") #TRUE
is_svo_correct(".") #TRUE
is_svo_correct("[]") #TRUE
}
