% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_uri}
\alias{is_uri}
\title{checks whether subject/verb/object is an item}
\usage{
is_uri(string)
}
\description{
checks whether subject/verb/object is an item
}
\examples{
glitter:::is_uri("elem") #returns FALSE
glitter:::is_uri("wd:Q456") #returns TRUE
glitter:::is_uri("<http://qsdfqsdfsqdf>") #returns TRUE
}
