% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_url}
\alias{is_url}
\title{checks whether element is a url ("<http_blah_>")}
\usage{
is_url(string)
}
\arguments{
\item{string}{a string}
}
\description{
checks whether element is a url ("<http_blah_>")
}
\examples{
is_url("<http://qsdqsdfqsdfqs.html>")
}
