% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_value}
\alias{is_value}
\title{checks whether element is a value ("'blah'" or '"blah"')}
\usage{
is_value(string)
}
\description{
checks whether element is a value ("'blah'" or '"blah"')
}
\examples{
glitter:::is_value("'blabla'") #TRUE
glitter:::is_value('"blabla"') #TRUE
glitter:::is_value('blabla') #FALSE
}
