% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_variable}
\alias{is_variable}
\title{Checks whether element is a variable ("?...")}
\usage{
is_variable(vstring)
}
\arguments{
\item{vstring}{a string or vector of strings}
}
\description{
Checks whether element is a variable ("?...")
}
\examples{
is_variable(c("?elem","?item")) #TRUE
}
