% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{isUpToDate}
\alias{isUpToDate}
\title{Is The Repository Up-To-Date?}
\usage{
isUpToDate(path_to_local_repo = NULL)
}
\arguments{
\item{path_to_local_repo}{path to local repo}
}
\description{
Is The Repository Up-To-Date?
}
