% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{isWorkingTreeClean}
\alias{isWorkingTreeClean}
\title{Check whether a commit is needed or not for a given repo}
\usage{
isWorkingTreeClean(path = getwd())
}
\arguments{
\item{path}{path to the local repo}
}
\description{
Check whether a commit is needed or not for a given repo
}
