% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{keep_prefix}
\alias{keep_prefix}
\title{keep only prefixes (decomposing paths if necessary)}
\usage{
keep_prefix(vstring)
}
\arguments{
\item{vstring}{a string or vector of strings}
}
\description{
keep only prefixes (decomposing paths if necessary)
}
\examples{
glitter:::keep_prefix("wdt:P31/wdt:P279*") # wdt
glitter:::keep_prefix("?item") # NULL
glitter:::keep_prefix("<http://qdsfqsdfqsfqsdf.html>") # NULL
glitter:::keep_prefix("wd:P343") # "wd:P343"
glitter:::keep_prefix("'11992081'^^xsd:integer") # xsd:integer
glitter:::keep_prefix(c("wd:P343","wdt:P249"))
}
