% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{keep_prefixed}
\alias{keep_prefixed}
\title{keep only prefixed elements and decompose paths}
\usage{
keep_prefixed(string)
}
\description{
keep only prefixed elements and decompose paths
}
\examples{
glitter:::keep_prefixed("wdt:P31/wdt:P279*") # wdt:P31 wdt:P279*
glitter:::keep_prefixed("?item") # NULL
glitter:::keep_prefixed("<http://qdsfqsdfqsfqsdf.html>") # NULL
glitter:::keep_prefixed("wd:P343") # "wd:P343"
glitter:::keep_prefixed("'11992081'^^xsd:integer") # xsd:integer
}
