% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tracking.R
\name{last_file_commit}
\alias{last_file_commit}
\title{Commit History of a File}
\usage{
last_file_commit(file, path = getwd())
}
\description{
This function can be especially useful in cases where a deleted files is being tracked down.
}
