% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_install_this_repo.R
\name{lfs_install_this_repo}
\alias{lfs_install_this_repo}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
lfs_install_this_repo(large_file_extensions = c("xlsx", "csv", "RData"))
}
\arguments{
\item{large_file_extensions}{extensions of files that will be large and will need to be tracked using the lfs}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
\keyword{internal}
