% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tracking.R
\name{list_file_commits}
\alias{list_file_commits}
\title{Commit History of a File}
\usage{
list_file_commits(file, path = getwd(), verbose = TRUE)
}
\description{
This function can be especially useful in cases where a deleted files is being tracked down.
}
