% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-tags.R, R/tags.R
\name{list_tags}
\alias{list_tags}
\title{FUNCTION_TITLE}
\usage{
list_tags(path = getwd())

list_tags(path = getwd())
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: getwd()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

`git2r::tags()` ordered by date.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[git2r]{tags}}
 \code{\link[purrr]{map}}
 \code{\link[tibble]{as_tibble}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{vars}}

\code{\link[git2r]{tags}}
}
