% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_unstaged_files}
\alias{list_unstaged_files}
\title{Get a list of unstaged files in a repo}
\usage{
list_unstaged_files(path = getwd(), verbose = TRUE)
}
\description{
Get a list of unstaged files in a repo
}
