% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_untracked_files}
\alias{list_untracked_files}
\title{Get a list of untracked files in a repo}
\usage{
list_untracked_files(path, verbose = TRUE)
}
\arguments{
\item{path}{path to local repo}
}
\value{
The complete git status message and untracked files are printed in the console, and a vector of the untracked filenames is invisibly returned.
}
\description{
This function takes the git status message and isolates the files that are new/untracked according to that message.
}
\keyword{internal}
