% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghiFunctions.R
\name{listOpenIssues}
\alias{listOpenIssues}
\title{List Open GitHub Issues}
\usage{
listOpenIssues(repo = NULL)
}
\description{
All GitHub Issues functions in this package requires downloading and setting up authorization for \href{https://github.com/stephencelis/ghi}{GHI}.
}
