% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lsFilesToCommit}
\alias{lsFilesToCommit}
\title{List Files To Commit}
\usage{
lsFilesToCommit(path_to_local_repo = NULL, label = FALSE)
}
\description{
This function combines all the Unstaged and Untracked files in a repo.
}
