% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsStagedFiles.R
\name{lsStagedFiles}
\alias{lsStagedFiles}
\title{List Staged Files}
\usage{
lsStagedFiles(
  path_to_local_repo = NULL,
  modified = TRUE,
  deleted = TRUE,
  new_file = TRUE,
  renamed = TRUE,
  label = FALSE
)
}
\arguments{
\item{label}{If TRUE, returns the results with prefix of "modified:|deleted:|new file: {file path}"}
}
\value{
If staged files exist, a vector of all the staged file paths.
}
\description{
This function lists all the staged files in a local repo.
}
