% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lsUnstagedFiles}
\alias{lsUnstagedFiles}
\title{List Unstaged Files}
\usage{
lsUnstagedFiles(
  path_to_local_repo = NULL,
  modified = TRUE,
  deleted = TRUE,
  label = FALSE
)
}
\value{
If unstaged files exist, a vector of all the untracked file paths.
}
\description{
This function lists all the unstaged files in a local repo.
}
\seealso{
\code{\link[purrr]{keep}}
 \code{\link[stringr]{str_replace}}
 \code{\link[secretary]{typewrite}}
}
