% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lsUntrackedFiles}
\alias{lsUntrackedFiles}
\title{List Untracked Files}
\usage{
lsUntrackedFiles(path_to_local_repo = NULL)
}
\value{
If untracked files exist, a vector of all the untracked file paths.
}
\description{
This function lists all the untracked files in a local repo.
}
