% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging-make.R
\name{makeDescriptionLinks}
\alias{makeDescriptionLinks}
\title{Make Links to Add to DESCRIPTION}
\usage{
makeDescriptionLinks(github_user, repo)
}
\description{
Make Links to Add to DESCRIPTION
}
\examples{
makeDescriptionLinks(github_user = "meerapatelmd",
                     repo = "rubix")
}
