% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_functions}
\alias{misc_functions}
\title{Correspondance between R-DSL functions and SPARQL misc functions.}
\format{
A data frame.
\describe{
\item{R}{R-DSL function}
\item{SPARQL}{SPARQL function}
}
}
\usage{
misc_functions
}
\description{
Correspondance between R-DSL functions and SPARQL misc functions.
}
\keyword{datasets}
