% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_files.R
\name{modified_files}
\alias{modified_files}
\title{Get a list of modified files in a repo}
\usage{
modified_files(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{path to local repo}
}
\description{
Get a list of modified files in a repo
}
