% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{openIssue}
\alias{openIssue}
\title{(Deprecated) Open a GitHub Issue}
\usage{
openIssue(path = getwd(), title, description)
}
\description{
(Deprecated) This function opens a GitHub Issue from the R console.
}
