% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{operators}
\alias{operators}
\title{Correspondance between R-DSL functions and SPARQL term functions.}
\format{
A data frame.
\describe{
\item{R}{R-DSL operator}
\item{SPARQL}{SPARQL operator}
}
}
\usage{
operators
}
\description{
Correspondance between R-DSL functions and SPARQL term functions.
}
\keyword{datasets}
