% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{parse_status_response}
\alias{parse_status_response}
\title{Parse Response to Git Status}
\usage{
parse_status_response(status_response)
}
\value{
A list of vectors that has split on the following headers: "On branch","Changes to be committed:", "Changes not staged for commit:", "Untracked files:" with headers removed.
}
\description{
Parse the status response received when calling \code{\link{status}} to filter for new, modified, deleted, staged, and unstaged files.
}
\seealso{
\code{\link[purrr]{keep}},\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
 \code{\link[rubix]{map_names_set}}
}
