% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{parseStatusMessage}
\alias{parseStatusMessage}
\title{Parse Status Message}
\usage{
parseStatusMessage(statusMessage)
}
\value{
A list of vectors that has split on the following headers: "On branch","Changes to be committed:", "Changes not staged for commit:", "Untracked files:" with headers removed.
}
\description{
This function parses the status message returned as a vector when calling the status() function
}
\seealso{
\code{\link[purrr]{keep}},\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
 \code{\link[rubix]{map_names_set}}
}
