% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty.R
\name{pretty}
\alias{pretty}
\title{Takes the command line message vector and prints a clean version in console.}
\usage{
pretty(glitter_output)
}
\arguments{
\item{glitter_output}{output vector from any glitter function}
}
\description{
Takes the command line message vector and prints a clean version in console.
}
