% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_git.R
\name{pretty_git}
\alias{pretty_git}
\title{Takes the command line message vector and prints a clean version in console. Deprecated for pretty() function.}
\usage{
pretty_git(glitter_output)
}
\arguments{
\item{glitter_output}{output vector from any glitter function}
}
\description{
Takes the command line message vector and prints a clean version in console. Deprecated for pretty() function.
}
\keyword{internal}
