% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_if_has_length.R
\name{print_if_has_length}
\alias{print_if_has_length}
\title{Prints the git message in console if the message is of length greater than zero}
\usage{
print_if_has_length(git_msg)
}
\arguments{
\item{git_msg}{output from a glitter git command}
}
\description{
Prints the git message in console if the message is of length greater than zero
}
