% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_if_exists.R
\name{print_if_message_exists}
\alias{print_if_message_exists}
\title{Prints the git message in console if the message is of length greater than zero}
\usage{
print_if_message_exists(git_msg)
}
\arguments{
\item{git_msg}{output from a glitter git command}
}
\description{
Prints the git message in console if the message is of length greater than zero
}
