% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_response}
\alias{print_response}
\title{Print the Command Line Response}
\usage{
print_response(git_response)
}
\arguments{
\item{git_response}{Output after a git command is sent to the command line.}
}
\description{
Print a command line response from a \code{\link{glitter}} function.
}
