% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printMsg}
\alias{printMsg}
\title{Print the Command Line Response}
\usage{
printMsg(git_msg)
}
\arguments{
\item{git_msg}{Output from a glitter function.}
}
\description{
This function prints any command line response from a glitter function if the response is of length greater than 0.
}
