% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull a Remote Repo}
\usage{
pull(path_to_local_repo = NULL, verbose = TRUE)
}
\arguments{
\item{path_to_local_repo}{full path to directory of the repo to be pulled}
}
\description{
Pull a Remote Repo
}
