% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_all_if_needed.R
\name{push_all_if_needed}
\alias{push_all_if_needed}
\title{Push a Local Repo if a Change is detected}
\usage{
push_all_if_needed(path_to_local_repo = NULL, commit_message)
}
\arguments{
\item{path_to_local_repo}{PARAM_DESCRIPTION, Default: NULL}

\item{commit_message}{PARAM_DESCRIPTION}
}
\description{
FUNCTION_DESCRIPTION
}
