% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_tag.R
\name{push_tag}
\alias{push_tag}
\title{Push a local repo to a GitHub repository}
\usage{
push_tag(
  path = getwd(),
  remote_name = "origin",
  remote_branch = "master",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{path}{full path to local repository to be pushed}

\item{remote_name}{name of remote to push to. Defaults to
"origin".}

\item{remote_branch}{name of branch on the remote to push
to. Defaults to "master" or "main" if the cardinal branch is using
this newer default.}
}
\description{
Push a local repo to a GitHub repository
}
