% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_this_repo.R
\name{push_this_repo}
\alias{push_this_repo}
\title{Push a local repo to a GitHub repository}
\usage{
push_this_repo(remote_name = "origin", remote_branch = "master")
}
\arguments{
\item{remote_name}{name of remote to push to. Defaults to "origin".}

\item{remote_branch}{name of branch on the remote to push to. Defaults to "master".}
}
\description{
Push a local repo to a GitHub repository
}
