% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_wd.R
\name{push_wd}
\alias{push_wd}
\title{Push a local repo to a GitHub repository}
\usage{
push_wd(remote_name = "origin", remote_branch = "master")
}
\arguments{
\item{remote_name}{name of remote to push to. Defaults to "origin".}

\item{remote_branch}{name of branch on the remote to push to. Defaults to "master".}

\item{path_to_local_repo}{full path to local repository to be pushed}
}
\description{
Push a local repo to a GitHub repository
}
