% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{read_latest_desc_version}
\alias{read_latest_desc_version}
\title{FUNCTION_TITLE}
\usage{
read_latest_desc_version(path = getwd())
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: getwd()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
