% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{read_news}
\alias{read_news}
\title{FUNCTION_TITLE}
\usage{
read_news(path = getwd())
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: getwd()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[usethis]{use_news_md}},\code{\link[usethis]{ui}}
 \code{\link[xfun]{read_utf8}}
 \code{\link[rlang]{seq2}}
}
