% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGitIgnore.R
\name{readGitIgnore}
\alias{readGitIgnore}
\title{Read .gitignore}
\usage{
readGitIgnore(path_to_local_repo = NULL)
}
\description{
This function reads the .gitignore file in the repo. If one is not present, it is created.
}
