% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmFromGitIgnore.R
\name{rmFromGitIgnore}
\alias{rmFromGitIgnore}
\title{Remove from .gitignore}
\usage{
rmFromGitIgnore(..., commit = TRUE, path_to_local_repo = NULL)
}
\description{
This function remoes an entry in a .gitignore file if there is one in the repo path.
}
