% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_fns_for_cli.R
\name{sanitize_fns_for_cli}
\alias{sanitize_fns_for_cli}
\title{Subsitute spaces and special characters with the backslash for a CLI command}
\usage{
sanitize_fns_for_cli(vector)
}
\description{
Subsitute spaces and special characters with the backslash for a CLI command
}
