% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_sparql.R
\name{send_sparql}
\alias{send_sparql}
\title{Send SPARQL query to endpoint and get tibble as a result}
\usage{
send_sparql(
  .query,
  endpoint = "Wikidata",
  user_agent = getOption("glitter.ua",
    "glitter R package (https://github.com/lvaudor/glitter)"),
  max_tries = getOption("glitter.max_tries", 3L),
  max_seconds = getOption("glitter.max_seconds", 120L),
  timeout = getOption("glitter.timeout", 1000L),
  request_type = c("url", "body-form")
)
}
\arguments{
\item{.query}{a string corresponding to a SPARQL query}

\item{endpoint}{a string or url corresponding to a SPARQL endpoint. Defaults to "Wikidata"}

\item{user_agent}{a string indicating the user agent to send with the query.}

\item{max_tries, max_seconds}{Cap the maximum number of attempts with
\code{max_tries} or the total elapsed time from the first request with
\code{max_seconds}. If neither option is supplied (the default), \code{\link[httr2:req_perform]{req_perform()}}
will not retry.}

\item{timeout}{maximum number of seconds to wait (\code{httr2::req_timeout()}).}

\item{request_type}{a string indicating how the query should be sent: in the
URL (\code{url}, default, most common) or as a body form (\code{body-form}).}
}
\description{
Send SPARQL query to endpoint and get tibble as a result
}
\examples{
metro_query='SELECT ?item ?itemLabel ?coords
{
 ?item wdt:P361 wd:Q1552;
 wdt:P625 ?coords.
 OPTIONAL{?item wdt:P1619 ?date.}
 SERVICE wikibase:label { bd:serviceParam wikibase:language "en" . }
} ORDER BY ?itemLabel
'
send_sparql(metro_query)
}
