% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{set_functions}
\alias{set_functions}
\title{Correspondance between R-DSL functions and SPARQL set functions.}
\format{
A data frame.
\describe{
\item{R}{R-DSL function}
\item{SPARQL}{SPARQL set function}
\item{args}{list-column with R vs SPARQL argument names}
}
}
\usage{
set_functions
}
\description{
Correspondance between R-DSL functions and SPARQL set functions.
}
\keyword{datasets}
