% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_classes.R
\name{show_graph_classes}
\alias{show_graph_classes}
\title{Show the graph of sub/superclasses}
\usage{
show_graph_classes(tib_g, n_min = 10, layout = "kk")
}
\arguments{
\item{tib_g}{tib_g}

\item{n_min}{minimal number}

\item{layout}{layout of the graph for instance "kk" or "sugiyama"}
}
\description{
Show the graph of sub/superclasses
}
\examples{
tib_g=build_graph_classes("wd:Q627272")
show_graph_classes(tib_g,layout="kk")
show_graph_classes(tib_g,layout="sugiyama")
}
