% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{set_functions}
\alias{set_functions}
\alias{term_functions}
\alias{misc_functions}
\alias{string_functions}
\alias{numeric_functions}
\alias{datetime_functions}
\alias{operators}
\alias{all_correspondences}
\title{Correspondence between R-DSL functions and SPARQL functions/operators.}
\format{
A data frame.
\describe{
\item{R}{R-DSL function}
\item{SPARQL}{SPARQL function}
\item{args}{list-column with R vs SPARQL argument names}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58 rows and 2 columns.
}
\usage{
set_functions

term_functions

misc_functions

string_functions

numeric_functions

datetime_functions

operators

all_correspondences
}
\description{
Correspondence between R-DSL functions and SPARQL functions/operators.
}
\keyword{datasets}
