% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql-protection.R
\name{spq}
\alias{spq}
\alias{is.spq}
\alias{as.spq}
\title{SPARQL escaping.}
\usage{
spq(...)

is.spq(x)

as.spq(x)
}
\arguments{
\item{...}{Character vectors that will be combined into a single SPARQL
expression.}

\item{x}{Object to coerce}
}
\description{
Like \code{dbplyr::spq()}.
}
