% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_arrange.R
\name{spq_arrange}
\alias{spq_arrange}
\title{Arrange results by variable value}
\usage{
spq_arrange(query, vars)
}
\arguments{
\item{query}{a list with elements of the query}

\item{vars}{variables by which to arrange}
}
\description{
Arrange results by variable value
}
\examples{
tib=spq_init() \%>\%
spq_add("?item wdt:P31 wd:Q5",label="?item") \%>\%
spq_add("?item wdt:P19/wdt:P131* wd:Q60") \%>\%
spq_add("?item wikibase:sitelinks ?sitelinks") \%>\%
spq_arrange(c("DESC(?sitelinks)")) \%>\%
spq_head(n=3) \%>\%
spq_perform()
}
