% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sparql.R
\name{spq_assemble}
\alias{spq_assemble}
\title{Assemble query parts into a proper SPARQL query}
\usage{
spq_assemble(.query, endpoint = "Wikidata")
}
\arguments{
\item{.query}{a list with elements of the query}

\item{endpoint}{SPARQL endpoint to send the query to}
}
\description{
Assemble query parts into a proper SPARQL query
}
\examples{
spq_init() \%>\%
spq_add("?city wdt:P31 wd:Q515", .label="?city") \%>\%
spq_add("?city wdt:P1082 ?pop") \%>\%
spq_language("fr") \%>\%
spq_head(n=5) \%>\%
spq_assemble() \%>\%
cat()
}
