% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_filter.R
\name{spq_filter}
\alias{spq_filter}
\title{Builds the "filter" part of a query.}
\usage{
spq_filter(query = NULL, expressions)
}
\arguments{
\item{query}{a list with elements of the query*}

\item{expressions}{the vector of filtering expressions to apply}
}
\description{
Builds the "filter" part of a query.
}
\examples{
# Corpus of articles with "wikidata" in the title
spq_init() \%>\%
add_triplets("?item wdt:P31 wd:Q13442814") \%>\%
add_triplets("?item rdfs:label ?itemTitle") \%>\%
spq_filter("CONTAINS(LCASE(?itemTitle),'wikidata')") \%>\%
spq_filter("LANG(?itemTitle)='en'") \%>\%
spq_head(n=5) \%>\%
send()
}
