% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_head.R
\name{spq_head}
\alias{spq_head}
\title{Builds the "limit" part of a query.}
\usage{
spq_head(query, n = 5)
}
\arguments{
\item{query}{a list with elements of the query}

\item{n}{the maximum number of lines to return (defaults to 5)}
}
\description{
Builds the "limit" part of a query.
}
\examples{
add_triplets(s="?item",v="wdt:P31",o="wd:Q5",label="?item") \%>\%
add_triplets(s="?item",v="wdt:P19/wdt:P131*",o="wd:Q60") \%>\%
add_triplets(s="?item",v="wikibase:sitelinks",o="?sitelinks") \%>\%
spq_arrange(c("DESC(?sitelinks)")) \%>\%
spq_head(n=3) \%>\%
send()
}
