% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_label.R
\name{spq_label}
\alias{spq_label}
\title{Label variables}
\usage{
spq_label(.query, ..., .languages = getOption("glitter.lang", "en"))
}
\arguments{
\item{.query}{a list with elements of the query}

\item{...}{variables by which to arrange
(or SPARQL strings escaped with \code{spq()}, or strings, see examples)}

\item{.languages}{}
}
\description{
Label variables
}
