% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_language.R
\name{spq_language}
\alias{spq_language}
\title{Get labels in a specified language (apply only to Wikidata)}
\usage{
spq_language(query = NULL, language = "en")
}
\arguments{
\item{query}{the query}

\item{language}{the language in which the labels will be provided (defaults to "en" for English)}
}
\description{
Get labels in a specified language (apply only to Wikidata)
}
\examples{
tib=spq_init() \%>\%
 spq_add("?film wdt:P31 wd:Q11424", label="?film") \%>\%
 spq_add("?film wdt:P840 ?loc", label="?loc") \%>\%
 spq_add("?film wdt:P495 ?origin", label="?origin") \%>\%
 spq_add("?film wdt:P577 ?date") \%>\%
 spq_mutate(c("?year"="year(?date)")) \%>\%
 spq_language("fr") \%>\%
 spq_head(10) \%>\%
 spq_select("-?date") \%>\%
 send()
}
