% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_offset.R
\name{spq_offset}
\alias{spq_offset}
\title{Offset the first generated result}
\usage{
spq_offset(query, n = 5)
}
\arguments{
\item{query}{a list with elements of the query}

\item{n}{the maximum number of lines to return}
}
\description{
Offset the first generated result
}
\section{Subsetting}{

\code{\link[=spq_offset]{spq_offset()}} and \code{\link[=spq_head]{spq_head()}} are only useful when used with
\code{\link[=spq_arrange]{spq_arrange()}} that makes the order of results predictable.
}

\examples{
# Return 42 items
spq_init() \%>\%
spq_add("?item wdt:P31 wd:Q5",label="?item") \%>\%
spq_add("?item wdt:P19/wdt:P131* wd:Q60") \%>\%
spq_add("?item wikibase:sitelinks ?linkcount") \%>\%
spq_arrange(desc(linkcount)) \%>\%
spq_head(n=42)

# Return 42 items after the first 11 items
spq_init() \%>\%
spq_add("?item wdt:P31 wd:Q5",label="?item") \%>\%
spq_add("?item wdt:P19/wdt:P131* wd:Q60") \%>\%
spq_add("?item wikibase:sitelinks ?linkcount") \%>\%
spq_arrange(desc(linkcount)) \%>\%
spq_head(42) \%>\%
spq_offset(11)
}
