% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_prefix.R
\name{spq_prefix}
\alias{spq_prefix}
\title{Builds the "prefixes" part of a query.}
\usage{
spq_prefix(query = NULL, prefixes = NULL)
}
\arguments{
\item{query}{a list with elements of the query}

\item{prefixes}{a vector of prefixes}
}
\description{
Builds the "prefixes" part of a query.
}
