% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_select.R
\name{spq_select}
\alias{spq_select}
\title{Select (and create) particular variables}
\usage{
spq_select(query = NULL, ...)
}
\arguments{
\item{query}{a list with elements of the query}

\item{...}{variables by which to arrange (or SPARQL strings escaped with \code{spq()}, see examples)}
}
\description{
Select (and create) particular variables
}
\examples{

query = spq_init()
spq_select(query, count = n (human), eyecolorLabel, haircolorLabel)
}
