% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_set.R
\name{spq_set}
\alias{spq_set}
\title{Set helper values for the query}
\usage{
spq_set(.query, ...)
}
\arguments{
\item{.query}{query}

\item{...}{Helper values and their definition.}
}
\value{

}
\description{
Set helper values for the query (helps with readability)
}
\examples{
# find the individuals of the species
spq_init() \%>\%
# dog, cat or chicken
spq_set(species = c('wd:144','wd:146', 'wd:780'), mayor = "Q30185") \%>\%
spq_add(mayor = wdt::P31(species)) \%>\%
spq_add(.triple_pattern = "?mayor p:P39 ?node") \%>\%
# of mayor
spq_add(.triple_pattern = "?node ps:P39 wd:Q30185") \%>\%
# of some places
spq_add(.triple_pattern = "?node pq:P642 ?place") \%>\%
spq_perform()
}
