% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_summarise.R
\name{spq_summarise}
\alias{spq_summarise}
\title{Builds the "select" part of a query.}
\usage{
spq_summarise(query, vars)
}
\arguments{
\item{query}{a list with elements of the query}

\item{vars}{a vector with the names of the variables used for grouping}
}
\description{
Builds the "select" part of a query.
}
\examples{
add_triplets(s="?item",v="wdt:P361",o="wd:Q297853") \%>\%
add_triplets(s="?item",v="wdt:P1082",o="?folkm_ngd") \%>\%
add_triplets(s="?area",v="wdt:P31",o="wd:Q1907114",label="?area") \%>\%
add_triplets(s="?area",v="wdt:P527",o="?item") \%>\%
spq_group_by(c("?area","?areaLabel")) \%>\%
spq_summarise(c("?total_folkm"="sum(?folkm_ngd)")) \%>\%
send()
}
