% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOftest.R, R/status.R, R/test.R
\name{status}
\alias{status}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
status(path_to_local_repo)

status(path_to_local_repo)

status(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}

\item{remote_name}{name of remote to push to. Defaults to "origin".}

\item{remote_branch}{name of branch on the remote to push to. Defaults to "master".}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository

Push a local repo to remote MSK KMI Enterprise GitHub repository

Push a local repo to remote MSK KMI Enterprise GitHub repository
}
