% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_outside_repo.R
\name{status_outside_repo}
\alias{status_outside_repo}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
status_outside_repo(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}

\item{remote_name}{name of remote to push to. Defaults to "origin".}

\item{remote_branch}{name of branch on the remote to push to. Defaults to "master".}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
