% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_this_repo.R
\name{status_this_repo}
\alias{status_this_repo}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
status_this_repo()
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
