% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_if_dir_not_exist.R
\name{stop_if_dir_not_exist}
\alias{stop_if_dir_not_exist}
\title{Stop if the directory does not exist}
\usage{
stop_if_dir_not_exist(path_to_local_repo = NULL)
}
\description{
Stop if the directory does not exist
}
