% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{term_functions}
\alias{term_functions}
\title{Correspondance between R-DSL functions and SPARQL term functions.}
\format{
A data frame.
\describe{
\item{R}{R-DSL function}
\item{SPARQL}{SPARQL set function}
}
}
\usage{
term_functions
}
\description{
Correspondance between R-DSL functions and SPARQL term functions.
}
\keyword{datasets}
