% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{transform_wikidata_coords}
\alias{transform_wikidata_coords}
\title{Get coords at lat-lng from Wikidata format}
\usage{
transform_wikidata_coords(data, coord_column, prefix = "")
}
\arguments{
\item{data}{a tibble}

\item{coord_column}{the name of the column with coordinates formatted as 'Point(..... ....)'}

\item{prefix}{the prefix to be appended to "lat-lng" columns (defaults to "" -no prefix-)}
}
\description{
Get coords at lat-lng from Wikidata format
}
\examples{
\dontrun{
spq_init() \%>\%
  spq_add("?city wdt:P31/wdt:P279* wd:Q486972") \%>\%
  spq_label(city) \%>\%
  spq_mutate(coords = wdt::P625(city),
          .within_distance=list(center=c(long=4.84,lat=45.76),
                               radius=5)) \%>\%
  spq_perform() \%>\%
  transform_wikidata_coords("coords")
}
}
