% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untracked_files.R
\name{untracked_files}
\alias{untracked_files}
\title{Get a list of untracked files in a repo}
\usage{
untracked_files(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{path to local repo}
}
\description{
Get a list of untracked files in a repo
}
