% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untracked_files_in_wd.R
\name{untracked_files_in_wd}
\alias{untracked_files_in_wd}
\title{Get all untracked files in the current working dir}
\usage{
untracked_files_in_wd()
}
\description{
Get all untracked files in the current working dir
}
