% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usual_endpoints}
\alias{usual_endpoints}
\title{Usual endpoints: this dataset allows the user to refer to them using a simplified name rather than their full url.}
\format{
A data frame with usual SPARQL endpoints and abbreviated names
\describe{
\item{name}{the abbreviated name of the SPARQL endpoint}
\item{url}{the full address of the SPARQL endpoint}
...
}
}
\usage{
usual_endpoints
}
\description{
Usual endpoints: this dataset allows the user to refer to them using a simplified name rather than their full url.
}
\keyword{datasets}
