% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usual_prefixes}
\alias{usual_prefixes}
\title{Usual prefixes: this dataset allows the user to refer to usual prefixes in their queries without manually specifying the associated urls.}
\format{
A data frame with usual prefixes
\describe{
\item{type}{the type of prefix}
\item{name}{the prefix itself}
\item{url}{the corresponding ontology}
...
}
}
\usage{
usual_prefixes
}
\description{
Usual prefixes: this dataset allows the user to refer to usual prefixes in their queries without manually specifying the associated urls.
}
\keyword{datasets}
