% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{validate_versions}
\alias{validate_versions}
\title{FUNCTION_TITLE}
\usage{
validate_versions()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
