% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_add_all.R
\name{wd_add_all}
\alias{wd_add_all}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
wd_add_all()
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
