% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_add_commit_all.R
\name{wd_add_commit_all}
\alias{wd_add_commit_all}
\title{Add and commit all modified and untracked files in the current working directory}
\usage{
wd_add_commit_all(commit_message = NULL, description = NULL)
}
\arguments{
\item{commit_message}{defaults to NULL, where the commit_message is a "modify/add {filename} in {path to present R script}"}
}
\description{
Add and commit all modified and untracked files in the current working directory
}
