% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_add_commit_some.R
\name{wd_add_commit_some}
\alias{wd_add_commit_some}
\title{Add and commit some files in the current working directory}
\usage{
wd_add_commit_some(filenames, commit_message = NULL, description = NULL)
}
\arguments{
\item{commit_message}{defaults to NULL, where the commit_message is a "modify/add {filename} in {path to present R script}"}
}
\description{
Add and commit some files in the current working directory
}
