% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wd_properties}
\alias{wd_properties}
\title{Wikidata properties}
\format{
A data frame with 8939 rows and 5 variables:
\describe{
  \item{id}{id}
  \item{type}{property type}
  \item{label}{property label}
  \item{description}{property description}
  \item{altLabel}{alternative labels}
  ...
}
}
\source{
Wikidata \url{`r wikidata_url()`}
}
\usage{
wd_properties
}
\description{
Wikidata properties
}
\keyword{datasets}
