% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_status.R
\name{wd_status}
\alias{wd_status}
\title{Get the Git status of the current working directory}
\usage{
wd_status()
}
\value{
If the git message is of a length greater than 0, it is returned as a character vector and also printed in the console
}
\description{
This function differs from status() by the fact that the system command does not set a path to a given local repo, but uses the path that is the working directory.
}
