% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ac}
\alias{ac}
\title{Add and Commit}
\usage{
ac(
  ...,
  path = getwd(),
  pattern = NULL,
  commit_msg,
  all.files = FALSE,
  recursive = FALSE,
  ignore.case = FALSE,
  include.dirs = FALSE,
  no.. = FALSE,
  max_mb = 50,
  verbose = TRUE
)
}
\description{
Add and Commit
}
