% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{add}
\alias{add}
\title{Add Files}
\usage{
add(
  ...,
  path = getwd(),
  pattern = NULL,
  all.files = FALSE,
  recursive = FALSE,
  ignore.case = FALSE,
  include.dirs = FALSE,
  no.. = FALSE,
  max_mb = 50
)
}
\description{
Add Files
}
\seealso{
\code{\link[cli]{cat_line}}
 \code{\link[secretary]{character(0)}},\code{\link[secretary]{press_enter}}
 \code{\link[rlang]{list2}}
 \code{\link[purrr]{map}}
}
