% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commit.R, R/add_commit_some.R
\name{add_commit_some}
\alias{add_commit_some}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
add_commit_some(
  path_to_local_repo,
  filenames,
  commit_message = NULL,
  description = NULL
)

add_commit_some(
  path_to_local_repo,
  filenames,
  commit_message = NULL,
  description = NULL
)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}

\item{commit_message}{defaults to NULL, where the commit_message is a "modify/add {filename} in {path to present R script}"}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository

Push a local repo to remote GitHub repository
}
