% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commit_some.R
\name{add_commit_some}
\alias{add_commit_some}
\title{Commit some, but not all files, in a local repo to remote GitHub repository}
\usage{
add_commit_some(
  path_to_local_repo,
  filenames,
  commit_message = NULL,
  description = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}

\item{filenames}{names of files in the local repository path to be committed}

\item{commit_message}{defaults to NULL, where the commit_message is a "change to {filename} in {path to present R script}"}
}
\description{
Commit some, but not all files, in a local repo to remote GitHub repository
}
