% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_some.R
\name{add_some}
\alias{add_some}
\title{Add Some Files to a Commit}
\usage{
add_some(filenames, verbose = TRUE, path_to_local_repo = NULL)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}
}
\description{
This function takes a vector of filenames, filters them for only the untracked and unstaged files, and adds them to a commit.
}
