% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_some.R
\name{add_some}
\alias{add_some}
\title{Push a local repo to remote MSK KMI Enterprise GitHub repository}
\usage{
add_some(path_to_local_repo = NULL, filenames)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}
}
\description{
Push a local repo to remote MSK KMI Enterprise GitHub repository
}
