% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_query.R
\name{add_triplets}
\alias{add_triplets}
\title{Compose query parts for triplets with subject verb object.}
\usage{
add_triplets(
  query = NA,
  subject,
  verb,
  object,
  optional = FALSE,
  label = NA,
  limit = NA,
  within_box = c(NA, NA),
  within_distance = c(NA, NA)
)
}
\arguments{
\item{subject}{an anonymous variable (for instance, and by default, "?subject") or item (for instance "Q456"))}

\item{verb}{the property (for instance "wdt:P190")}

\item{object}{an anonymous variable (for instance, and by default, "?object") or item (for instance "Q456"))}

\item{optional}{whether the triplet is optional (other results are sent back even if this particular triplet is missing)}

\item{label}{a vector of variables for which to include a label column (defaults to NA)}

\item{limit}{the max number of items sent back}

\item{within_box}{if provided, rectangular bounding box for the triplet query.
Provided as list(southwest=c(long=...,lat=...),northeast=c(long=...,lat=...))}

\item{within_distance}{if provided, circular bounding box for the triplet query.
Provided as list(center=c(long=...,lat=...), radius=...), with radius in kilometers.
The center can also be provided as a variable (for instance, "?location") for the center coordinates to be retrieved directly from the query.}
}
\description{
Compose query parts for triplets with subject verb object.
}
\examples{
add_triplets(subject="?city",verb="wdt:P31/wdt:P279*",object="wd:Q515", label=c("?city"), limit=10) \%>\%
add_triplets(subject="?city",verb="wdt:P1082",object="?pop",optional=TRUE) \%>\%
add_triplets(subject="?city",verb="wdt:P625",object="?coords",in_box=list(southwest=c(3,43),northeast=c(7,47))) \%>\%
build_sparql() \%>\%
send_sparql()
}
