% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_github_functions.R
\name{browse_github_functions}
\alias{browse_github_functions}
\title{Browse GitHub Functions}
\arguments{
\item{github_username}{GitHub username}

\item{repo}{Repository belonging to the GitHub user}
}
\description{
These functions open up GitHub urls related to a repo in the browser returned by \code{\link{get_remote_urls}}.
}
