% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{browse_repo2}
\alias{browse_repo2}
\title{Open a GitHub Repo Issues Page in a Browser}
\usage{
browse_repo2(github_username, repo)
}
\description{
Open a GitHub Repo Issues Page in a Browser
}
\seealso{
\code{\link[utils]{browseURL}}
}
