% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPushSite.R, R/build_push_site.R
\name{buildPushSite}
\alias{buildPushSite}
\alias{build_push_site}
\title{Build and Push docs/ dir For GitHub Pages}
\usage{
buildPushSite(
  commit_message = "update GitHub Page",
  update_gitignore = TRUE,
  lazy = TRUE,
  preview = FALSE,
  devel = TRUE,
  ...
)

build_push_site(
  commit_message = "update GitHub Page",
  update_gitignore = TRUE,
  lazy = TRUE,
  preview = FALSE,
  devel = TRUE,
  ...
)
}
\description{
This function automatically documents, pushes, and installs a package, assuming that the basename fo the working directory is the same as the repo as in patelm9/{repo}. If the URL of the GitHub remote belongs to an Enterprise GitHub Repository, the package is instead installed using a Git hyperlink.

This function automatically documents, pushes, and installs a package, assuming that the basename fo the working directory is the same as the repo as in patelm9/{repo}. If the URL of the GitHub remote belongs to an Enterprise GitHub Repository, the package is instead installed using a Git hyperlink.
}
\seealso{
\code{\link[usethis]{use_pkgdown}}
 \code{\link[pkgdown]{build_site}}

\code{\link[usethis]{use_pkgdown}}
 \code{\link[pkgdown]{build_site}}
}
\keyword{internal}
