% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_push_site.R
\name{build_push_site}
\alias{build_push_site}
\title{Re-Build and Push GitHub Pages}
\usage{
build_push_site(
  commit_message = "update GitHub Page",
  lazy = TRUE,
  preview = FALSE,
  devel = TRUE,
  ...
)
}
\description{
This unlinks the docs subdirectory, runs devtools document function, and writes a new docs/ directory using the pkgdown build_site function.
}
\seealso{
\code{\link[usethis]{use_pkgdown}}
 \code{\link[pkgdown]{build_site}}
}
