% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_query.R
\name{build_sparql}
\alias{build_sparql}
\title{Assemble query parts into a query.}
\usage{
build_sparql(query_parts)
}
\arguments{
\item{query_parts}{a list with elements of the query}
}
\description{
Assemble query parts into a query.
}
\examples{
add_triplets(subject="?city",verb="wdt:P31",object="wd:Q515", limit=10) \%>\%
add_triplets(subject="?city",verb="wdt:P1082",object="?pop", label="?city", language="en") \%>\%
build_sparql() \%>\%
send_sparql()
}
