% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_pull_needed.R
\name{check_if_pull_needed}
\alias{check_if_pull_needed}
\title{CHeck if pull for a given repo is needed}
\usage{
check_if_pull_needed(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{path to local repo}
}
\description{
CHeck if pull for a given repo is needed
}
