% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{clone}
\alias{clone}
\title{Clone a Git Repository}
\usage{
clone(github_user, repo, destination_path)
}
\arguments{
\item{destination_path}{Path of the destination directory to which the repo will be cloned.}
}
\value{
Cloned repo in the path of {destination_path/repo name} if the directory does not exist. Otherwise an error is thrown.
}
\description{
Clone a Git Repository
}
