% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit}
\alias{commit}
\title{Perform a Git Commit}
\usage{
commit(
  commit_message,
  description = NULL,
  verbose = TRUE,
  path_to_local_repo = NULL
)
}
\arguments{
\item{commit_message}{message to be included in the commit}

\item{path_to_local_repo}{full path to local repository to be pushed}
}
\description{
Perform a Git Commit
}
