% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore.R
\name{dedupeGitIgnore}
\alias{dedupeGitIgnore}
\title{Remove duplicates from .gitignore}
\usage{
dedupeGitIgnore(commit = TRUE, path_to_local_repo = NULL)
}
\description{
This function reads the .gitignore file, removes duplicates, and overwrites the gitignore file.
}
\seealso{
\code{\link[readr]{read_lines}}
}
\keyword{internal}
