% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{delete_branch}
\alias{delete_branch}
\title{Delete a Branch}
\usage{
delete_branch(path = getwd(), branch, force = FALSE)
}
\arguments{
\item{branch}{Name of branch to delete}

\item{force}{If TRUE, a `-D` delete is performed in lieu of `-d`}
}
\description{
Delete a branch.
}
