% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_all}
\alias{deploy_all}
\alias{deploy_gh_pages}
\title{Update Package and GitHub Pages}
\usage{
deploy_all(path = getwd(), lazy = TRUE, preview = FALSE, devel = TRUE, ...)

deploy_gh_pages(
  path = getwd(),
  lazy = TRUE,
  preview = FALSE,
  devel = TRUE,
  ...
)
}
\description{
This unlinks the docs subdirectory, runs devtools document function, and writes a new docs/ directory using the pkgdown build_site function.

This unlinks the docs subdirectory, runs devtools document function, and writes a new docs/ directory using the pkgdown build_site function.
}
\seealso{
\code{\link[usethis]{use_pkgdown}}
 \code{\link[pkgdown]{build_site}}

\code{\link[usethis]{use_pkgdown}}
 \code{\link[pkgdown]{build_site}}
}
