% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_gh_pages}
\alias{deploy_gh_pages}
\title{Re-Build and Push GitHub Pages}
\usage{
deploy_gh_pages(
  commit_msg = "deploy GitHub Pages using glitter",
  remote_name = "origin",
  remote_branch = "master",
  path = getwd(),
  examples = TRUE,
  run_dont_run = FALSE,
  seed = 1014,
  lazy = TRUE,
  override = list(),
  preview = FALSE,
  devel = TRUE,
  new_process = !devel,
  install = !devel
)
}
\arguments{
\item{commit_msg}{commit message}

\item{remote_name}{name of remote to push to. Defaults to
"origin".}

\item{remote_branch}{name of branch on the remote to push
to. Defaults to "master" or "main" if the cardinal branch is using
this newer default.}

\item{path}{full path to local repository to be pushed}

\item{examples}{Run examples?}

\item{run_dont_run}{Run examples that are surrounded in \\dontrun?}

\item{seed}{Seed used to initialize so that random examples are
reproducible.}

\item{lazy}{If \code{TRUE}, will only rebuild articles and reference pages
if the source is newer than the destination.}

\item{override}{An optional named list used to temporarily override
values in \verb{_pkgdown.yml}}

\item{preview}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}

\item{devel}{Use development or deployment process?

If \code{TRUE}, uses lighter-weight process suitable for rapid
iteration; it will run examples and vignettes in the current process,
and will load code with \code{pkgload::load_all()}.

If \code{FALSE}, will first install the package to a temporary library,
and will run all examples and vignettes in a new process.

\code{build_site()} defaults to \code{devel = FALSE} so that you get high fidelity
outputs when you building the complete site; \code{build_reference()},
\code{build_home()} and friends default to \code{devel = TRUE} so that you can
rapidly iterate during development.}

\item{new_process}{If \code{TRUE}, will run \code{build_site()} in a separate process.
This enhances reproducibility by ensuring nothing that you have loaded
in the current process affects the build process.}

\item{install}{If TRUE, installs the package after the changes are pushed to the remote, Default: TRUE}
}
\description{
This unlinks the docs subdirectory, runs devtools document function, and writes a new docs/ directory using the pkgdown build_site function.
}
\seealso{
\code{\link[usethis]{use_pkgdown}}
 \code{\link[pkgdown]{build_site}}
}
