% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFunctions.R
\name{fileCommitHistory}
\alias{fileCommitHistory}
\title{Commit History of a File}
\usage{
fileCommitHistory(filePath, repo_path, verbose = TRUE)
}
\description{
This function can be especially useful in cases where a deleted files is being tracked down.
}
