% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_claims}
\alias{get_claims}
\title{Get claims regarding one Wikidata thing}
\usage{
get_claims(id, with_labels = FALSE)
}
\arguments{
\item{id}{a Wikidata ID, either of an item ("Qxxxxx") or of a property ("Pxxxxx"), or the item itself.}

\item{language}{language of description, defaults to English ("en")}
}
\description{
Get claims regarding one Wikidata thing
}
\examples{
thing=get_thing("Q431603")
get_claims(thing, with_labels=TRUE)
}
