% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get description of Wikidata thing}
\usage{
get_info(id, language = "en", with_labels = FALSE)
}
\arguments{
\item{id}{a Wikidata ID, either of an item ("Qxxxxx") or of a property ("Pxxxxx"), or the item itself.}

\item{language}{language of description, defaults to English ("en")}
}
\description{
Get description of Wikidata thing
}
\examples{
thing=get_thing("Q431603")
get_info(thing)
}
