% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{get_repo_info}
\alias{get_repo_info}
\title{Get Information about GitHub Repository}
\usage{
get_repo_info(github_user, repo, as_list = FALSE)
}
\arguments{
\item{github_user}{GitHub username. Enterprise GitHub is not supported.}

\item{repo}{Repository name as a string (case insensitive).}

\item{as_list}{If TRUE, the repository information is returned as a list object instead of a single row tibble.}
}
\value{
A tibble or if `as_list` is set to `TRUE`, a list object.
}
\description{
Filter the output of `get_repos()` for a specific repository by name.
}
