% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_triplets}
\alias{get_triplets}
\title{Get triplets with subject verb object.}
\usage{
get_triplets(
  subject = "?subject",
  verb = "?verb",
  object = "?object",
  optional = FALSE,
  label = NA,
  limit = NA,
  within_box = c(NA, NA),
  within_distance = c(NA, NA)
)
}
\arguments{
\item{subject}{an anonymous variable (for instance, and by default, "?subject") or item (for instance "Q456"))}

\item{verb}{the property (for instance "wdt:P190")}

\item{object}{an anonymous variable (for instance, and by default, "?object") or item (for instance "Q456"))}

\item{label}{a vector of variables for which to include a label column (defaults to NA)}

\item{limit}{the max number of items sent back}

\item{within_box}{if provided, rectangular bounding box for the triplet query. Provided as list(southwest=c(long=...,lat=...),northeast=c(long=...,lat=...))}

\item{within_distance}{if provided, circular bounding box for the triplet query. Provided as list(center=c(long=...,lat=...), radius=...), with radius in kilometers. The center can also be provided as a variable (for instance, "?location") for the center coordinates to be retrieved directly from the query.}
}
\description{
Get triplets with subject verb object.
}
\examples{
get_triplets(subject="?city",verb="wdt:P31/wdt:P279*",object="wd:Q515", label=c("?city"), limit=10)
}
