% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_triplets.R
\name{get_triplets}
\alias{get_triplets}
\title{Get triplets with subject verb object.}
\usage{
get_triplets(
  triplet = NULL,
  subject = NULL,
  verb = NULL,
  object = NULL,
  required = TRUE,
  label = NA,
  limit = NULL,
  within_box = c(NA, NA),
  within_distance = c(NA, NA)
)
}
\arguments{
\item{triplet}{triplet}

\item{subject}{an anonymous variable (for instance, and by default, "?subject") or item (for instance "wd:Q456"))}

\item{verb}{the property (for instance "wdt:P190")}

\item{object}{an anonymous variable (for instance, and by default, "?object") or item (for instance "wd:Q456"))}

\item{required}{whether to require this triplet (Boolean)}

\item{label}{a vector of variables for which to include a label column (defaults to NA)}

\item{limit}{the max number of items sent back}

\item{within_box}{if provided, rectangular bounding box for the triplet query. Provided as list(southwest=c(long=...,lat=...),northeast=c(long=...,lat=...))}

\item{within_distance}{if provided, circular bounding box for the triplet query. Provided as list(center=c(long=...,lat=...), radius=...), with radius in kilometers. The center can also be provided as a variable (for instance, "?location") for the center coordinates to be retrieved directly from the query.}
}
\description{
Get triplets with subject verb object.
}
\examples{
get_triplets(s="wd:Q456",v="wdt:P625",o="?coords")
get_triplets(t="wd:Q456 wdt:P625 ?coords")
}
