% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_triplets}
\alias{get_triplets}
\title{Get triplets with subject verb object.}
\usage{
get_triplets(
  subject = "?subject",
  verb = "?verb",
  object = "?object",
  limit = 10,
  send_query = TRUE,
  explicit = FALSE
)
}
\arguments{
\item{subject}{an anonymous variable (for instance, and by default, "?subject") or item (for instance "Q456"))}

\item{verb}{the property (for instance "wdt:P190")}

\item{object}{an anonymous variable (for instance, and by default, "?object") or item (for instance "Q456"))}

\item{explicit}{T or F, whether or not to include all non-anonymous elements in results table}
}
\description{
Get triplets with subject verb object.
}
\examples{
get_triplets(subject="?subject",verb="wdt:P17",object="wd:Q35", limit=10)
get_triplets(subject="wd:Q456", verb="wdt:P17", object="?country")
get_triplets(subject="wd:Q456", verb="?prop", object="wd:Q142")
get_triplets(subject="wd:Q456", verb="wdt:P1082", object="?pop")
purrr::map_df(paste0("wd:",c("Q90","Q456","Q23482")),get_triplets,verb="wdt:P1082",explicit=TRUE)
}
