% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{hasRemote}
\alias{hasRemote}
\title{Does the GitHub Repo Have a Remote?}
\usage{
hasRemote(path_to_local_repo = NULL, remote_name = "origin")
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}
}
\value{
TRUE if the path has a remote url.
}
\description{
Does the GitHub Repo Have a Remote?
}
