% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{interpret_svo}
\alias{interpret_svo}
\title{interprets if element is an R code inclusion of the type {...}}
\usage{
interpret_svo(string)
}
\arguments{
\item{string}{a string}
}
\description{
interprets if element is an R code inclusion of the type {...}
}
\examples{
obj1="0000000121012885"
glitter:::interpret_svo("{obj1}")
obj2=c("wd:Q432","wd:Q576")
glitter:::interpret_svo("{obj2}")
}
