% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_prefix_known}
\alias{is_prefix_known}
\title{Checks whether the element has a known prefix}
\usage{
is_prefix_known(string, prefixes, endpoint = "Wikidata")
}
\arguments{
\item{string}{a string}
}
\description{
Checks whether the element has a known prefix
}
\examples{
glitter:::is_prefix_known("wd:Q343",endpoint="Wikidata") # TRUE
#glitter:::is_prefix_known("blop:blabla",endpoint="other", prefixes=usual_prefixes) #returns error message
}
