% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_prefixed}
\alias{is_prefixed}
\title{Checks whether element is prefixed ("prefix::id")}
\usage{
is_prefixed(string)
}
\arguments{
\item{string}{a string}
}
\description{
Checks whether element is prefixed ("prefix::id")
}
\examples{
glitter:::is_prefixed("elem") #returns FALSE
glitter:::is_prefixed("wd:Q456") #returns TRUE
glitter:::is_prefixed("<http://qsdfqsdfsqdf>") #returns TRUE
}
