% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_svo_correct}
\alias{is_svo_correct}
\title{Checks whether subject/verb/object is stated correctly}
\usage{
is_svo_correct(string)
}
\arguments{
\item{string}{a string}
}
\description{
Checks whether subject/verb/object is stated correctly
}
\examples{
glitter:::is_svo_correct("elem") #FALSE
glitter::is_svo_correct("{choices}") #TRUE
glitter:::is_svo_correct("a") #TRUE
glitter:::is_svo_correct("is") #TRUE
glitter:::is_svo_correct("'0000000121012885'") #TRUE
glitter:::is_svo_correct(".") #TRUE
}
