% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_svo_correct}
\alias{is_svo_correct}
\title{checks whether subject/verb/object is stated correctly}
\usage{
is_svo_correct(string)
}
\description{
checks whether subject/verb/object is stated correctly
}
\examples{
glitter:::is_svo_correct("elem")
}
