% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_url}
\alias{is_url}
\title{Checks whether element is a url ("<http_blah_>")}
\usage{
is_url(vstring)
}
\arguments{
\item{vstring}{a string or vector of strings}
}
\description{
Checks whether element is a url ("<http_blah_>")
}
\examples{
is_url(c("<http://qsdqsdfqsdfqs.html>","<http blablabla>")) #TRUE
}
