% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{is_variable}
\alias{is_variable}
\title{Checks whether element is a variable ("?...")}
\usage{
is_variable(string)
}
\arguments{
\item{string}{a string}
}
\description{
Checks whether element is a variable ("?...")
}
\examples{
glitter:::is_svo_correct("elem")
}
