% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{isRepo}
\alias{isRepo}
\title{Is the path a git repo?}
\usage{
isRepo(path_to_local_repo = NULL)
}
\arguments{
\item{path_to_local_repo}{full path to local repository to be pushed}
}
\value{
TRUE if the path has an initialized .git.
}
\description{
Is the path a git repo?
}
