% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWorkingTreeClean.R
\name{isWorkingTreeClean}
\alias{isWorkingTreeClean}
\title{Check whether a commit is needed or not for a given repo}
\usage{
isWorkingTreeClean(path_to_local_repo = NULL)
}
\arguments{
\item{path_to_local_repo}{path to the local repo}
}
\description{
Check whether a commit is needed or not for a given repo
}
