% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/list.R
\name{list_modified_files}
\alias{list_modified_files}
\title{Get a list of modified files in a repo}
\usage{
list_modified_files(path = getwd())

list_modified_files(path = getwd())
}
\arguments{
\item{path}{path to local repo}
}
\value{
The complete git status message and modified files are printed in the console, and a vector of the modified filenames is invisibly returned.

The complete git status message and modified files are printed in the console, and a vector of the modified filenames is invisibly returned.
}
\description{
This function takes the git status message and isolates the files that have been modified according to that message.

This function takes the git status message and isolates the files that have been modified according to that message.
}
\keyword{internal}
