% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_untracked_files}
\alias{list_untracked_files}
\title{Get a list of untracked files in a repo}
\usage{
list_untracked_files(path = getwd(), verbose = TRUE)
}
\arguments{
\item{path}{path to local repo}
}
\description{
This function takes the git status message and isolates the files that are new/untracked according to that message.
}
\keyword{internal}
