% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghiFunctions.R, R/listOpenIssues.R
\name{listOpenIssues}
\alias{listOpenIssues}
\title{GitHub Issues}
\usage{
listOpenIssues(repo = NULL)

listOpenIssues(repo = NULL)
}
\description{
https://github.com/stephencelis/ghi

https://github.com/stephencelis/ghi
}
