% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lsStagedFiles}
\alias{lsStagedFiles}
\title{List Staged Files}
\usage{
lsStagedFiles(
  path_to_local_repo = NULL,
  modified = TRUE,
  deleted = TRUE,
  new_file = TRUE,
  renamed = TRUE,
  label = FALSE
)
}
\arguments{
\item{modified}{PARAM_DESCRIPTION, Default: TRUE}

\item{deleted}{PARAM_DESCRIPTION, Default: TRUE}

\item{new_file}{PARAM_DESCRIPTION, Default: TRUE}

\item{renamed}{PARAM_DESCRIPTION, Default: TRUE}

\item{label}{If TRUE, returns the results with prefix of "modified:|deleted:|new file: {file path}"}
}
\value{
If staged files exist, a vector of all the staged file paths.
}
\description{
This function lists all the staged files in a local repo.
}
\seealso{
\code{\link[purrr]{keep}},\code{\link[purrr]{map}}
 \code{\link[stringr]{str_replace}}
 \code{\link[secretary]{typewrite}}
}
