% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_files.R
\name{modified_files}
\alias{modified_files}
\title{Get a list of modified files in a repo}
\usage{
modified_files(path_to_local_repo = NULL)
}
\arguments{
\item{path_to_local_repo}{path to local repo}
}
\value{
The complete git status message and modified files are printed in the console, and a vector of the modified filenames is invisibly returned.
}
\description{
This function takes the git status message and isolates the files that have been modified according to that message.
}
