% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull a GitHub repo based on the path to local repo}
\usage{
pull(path_to_local_repo = NULL, verbose = TRUE)
}
\arguments{
\item{path_to_local_repo}{full path to directory of the repo to be pulled. If the path does not exist, it will be cloned to the parent directory.}
}
\value{
pulled or cloned repo at the given path_to_local_repo
}
\description{
Pull a GitHub repo based on the path to local repo
}
