% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send}
\alias{send}
\title{Assemble query parts into a sparql query and send it to endpoint to get a tibble as a result.}
\usage{
send(query, endpoint = "Wikidata")
}
\arguments{
\item{query}{a list with elements of the query}

\item{endpoint}{a string or url corresponding to a SPARQL endpoint. Defaults to "Wikidata"}
}
\description{
Assemble query parts into a sparql query and send it to endpoint to get a tibble as a result.
}
\examples{
spq_add(subject="?city",verb="wdt:P31",object="wd:Q515") \%>\%
spq_add(subject="?city",verb="wdt:P1082",object="?pop", label="?city", language="en") \%>\%
spq_head(n=5) \%>\%
send()
}
