% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_sparql.R
\name{send_sparql}
\alias{send_sparql}
\title{Send SPARQL query to endpoint and get tibble as a result}
\usage{
send_sparql(query, endpoint = "Wikidata")
}
\arguments{
\item{query}{a string corresponding to a SPARQL query}

\item{endpoint}{a string or url corresponding to a SPARQL endpoint. Defaults to "Wikidata"}
}
\description{
Send SPARQL query to endpoint and get tibble as a result
}
\examples{
metro_query='SELECT ?item ?itemLabel ?coords
{
 ?item wdt:P361 wd:Q1552;
 wdt:P625 ?coords.
 OPTIONAL{?item wdt:P1619 ?date.}
 SERVICE wikibase:label { bd:serviceParam wikibase:language "en" . }
} ORDER BY ?itemLabel
'
send_sparql(query=metro_query)
}
