% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql-translate.R
\name{spq}
\alias{spq}
\alias{is.spq}
\alias{as.spq}
\title{SPARQL escaping.}
\usage{
spq(...)

is.spq(x)

as.spq(x, con)
}
\arguments{
\item{...}{Character vectors that will be combined into a single SPARQL
expression.}

\item{x}{Object to coerce}

\item{con}{Needed when \code{x} is directly supplied from the user so that
schema specifications can be quoted using the correct identifiers.}
}
\description{
Like \code{dbplyr::spq()}.
}
