% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_arrange.R
\name{spq_arrange}
\alias{spq_arrange}
\title{Arrange results by variable value}
\usage{
spq_arrange(query, vars)
}
\arguments{
\item{query}{a list with elements of the query}

\item{vars}{variables by which to arrange}
}
\description{
Arrange results by variable value
}
\examples{
add_triplets(s="?item",v="wdt:P31",o="wd:Q5",label="?item") \%>\%
add_triplets(s="?item",v="wdt:P19/wdt:P131*",o="wd:Q60") \%>\%
add_triplets(s="?item",v="wikibase:sitelinks",o="?sitelinks") \%>\%
spq_arrange(c("DESC(?sitelinks)")) \%>\%
spq_head(n=3) \%>\%
send()
}
