% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_arrange.R
\name{spq_arrange}
\alias{spq_arrange}
\title{Arrange results by variable value}
\usage{
spq_arrange(query, ..., replace = FALSE)
}
\arguments{
\item{query}{a list with elements of the query}

\item{replace}{whether to replace the pre-existing arranging}

\item{vars}{variables by which to arrange}
}
\description{
Arrange results by variable value
}
\examples{
# descending length, ascending itemLabel, "R" syntax
spq_init() \%>\%
  spq_add("?item wdt:P31/wdt:P279* wd:Q4022", label = c("?item")) \%>\%
  spq_add("?item wdt:P2043 ?length") \%>\%
  spq_add("?item wdt:P625 ?location") \%>\%
  spq_arrange(desc(length), itemLabel) \%>\%
  spq_head(50)

# descending length, ascending itemLabel, SPARQL syntax
spq_init() \%>\%
  spq_add("?item wdt:P31/wdt:P279* wd:Q4022", label = c("?item")) \%>\%
  spq_add("?item wdt:P2043 ?length") \%>\%
  spq_add("?item wdt:P625 ?location") \%>\%
  spq_arrange("DESC(?length) ?itemLabel") \%>\%
  spq_head(50)

# descending xsd:integer(mort), R syntax
spq_init() \%>\%
  spq_add("?oeuvre dcterms:creator ?auteur") \%>\%
  spq_add("?auteur bio:death ?mort") \%>\%
  spq_add("?auteur foaf:familyName ?nom") \%>\%
  spq_filter("xsd:integer(?mort)<'1924'^^xsd:integer") \%>\%
  spq_group_by(c("?auteur","?nom","?mort")) \%>\%
  spq_arrange(desc(xsd:integer(mort)))

# descending xsd:integer(mort), SPARQL syntax
spq_init() \%>\%
  spq_add("?oeuvre dcterms:creator ?auteur") \%>\%
  spq_add("?auteur bio:death ?mort") \%>\%
  spq_add("?auteur foaf:familyName ?nom") \%>\%
  spq_filter("xsd:integer(?mort)<'1924'^^xsd:integer") \%>\%
  spq_group_by(c("?auteur","?nom","?mort")) \%>\%
  spq_arrange("DESC(xsd:integer(?mort))")
}
