% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_filter.R
\name{spq_filter}
\alias{spq_filter}
\title{Builds the "filter" part of a query.}
\usage{
spq_filter(query = NULL, ...)
}
\arguments{
\item{query}{a list with elements of the query}

\item{...}{variables by which to arrange
(or SPARQL strings escaped with \code{spq()}, or strings, see examples)}
}
\description{
Builds the "filter" part of a query.
}
\examples{
# Corpus of articles with "wikidata" in the title
spq_init() \%>\%
spq_add("?item wdt:P31 wd:Q13442814") \%>\%
spq_add("?item rdfs:label ?itemTitle") \%>\%
spq_filter(str_detect(str_to_lower(itemTitle), 'wikidata')) \%>\%
spq_filter("LANG(?itemTitle)='en'") \%>\%
spq_head(n = 5)
}
