% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_group_by.R
\name{spq_group_by}
\alias{spq_group_by}
\title{Builds the "select" part of a query.}
\usage{
spq_group_by(query, vars)
}
\arguments{
\item{query}{a list with elements of the query}

\item{vars}{a vector with the names of the variables used for grouping}
}
\description{
Builds the "select" part of a query.
}
\examples{
add_triplets(t="?s a ?class", limit=3) \%>\%
spq_group_by("?class") \%>\%
send()
}
