% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_group_by.R
\name{spq_group_by}
\alias{spq_group_by}
\title{Group the results by one or more variables}
\usage{
spq_group_by(query, vars)
}
\arguments{
\item{query}{a list with elements of the query}

\item{vars}{a vector with the names of the variables used for grouping}
}
\description{
Group the results by one or more variables
}
\examples{
\dontrun{
spq_init() \%>\%
spq_add("?s a ?class") \%>\%
spq_group_by("?class") \%>\%
spq_head(n=3) \%>\%
send()
}
}
