% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_group_by.R
\name{spq_group_by}
\alias{spq_group_by}
\title{Group the results by one or more variables}
\usage{
spq_group_by(.query, ...)
}
\arguments{
\item{.query}{query}

\item{...}{Either R-DSL or strings with variable names}
}
\value{
A query object
}
\description{
Group the results by one or more variables
}
\examples{
spq_init() \%>\%
spq_add("?item wdt:P361 wd:Q297853") \%>\%
spq_add("?item wdt:P1082 ?folkm_ngd") \%>\%
spq_add("?area wdt:P31 wd:Q1907114", .label = "?area") \%>\%
spq_add("?area wdt:P527 ?item") \%>\%
spq_group_by(area, areaLabel)  \%>\%
spq_summarise(total_folkm = sum(folkm_ngd))
}
