% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_head.R
\name{spq_head}
\alias{spq_head}
\title{Return the first lines of results}
\usage{
spq_head(query, n = 5)
}
\arguments{
\item{query}{a list with elements of the query}

\item{n}{the maximum number of lines to return (defaults to 5)}
}
\description{
Return the first lines of results
}
\examples{
spq_init() \%>\%
add_triplets("?item wdt:P31 wd:Q5",label="?item") \%>\%
add_triplets("?item wdt:P19/wdt:P131* wd:Q60") \%>\%
add_triplets("?item wikibase:sitelinks ?sitelinks") \%>\%
spq_arrange(c("DESC(?sitelinks)")) \%>\%
spq_head(n=3) \%>\%
send()
}
