% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_head.R
\name{spq_head}
\alias{spq_head}
\title{Return the first lines of results}
\usage{
spq_head(query, n = 5)
}
\arguments{
\item{query}{a list with elements of the query}

\item{n}{the maximum number of lines to return}
}
\description{
Return the first lines of results
}
\section{Subsetting}{

\code{\link[=spq_offset]{spq_offset()}} and \code{\link[=spq_head]{spq_head()}} are only useful when used with
\code{\link[=spq_arrange]{spq_arrange()}} that makes the order of results predictable.
}

\examples{
# Return the default of 5 items
spq_init() \%>\%
spq_add("?item wdt:P31 wd:Q5",label="?item") \%>\%
spq_add("?item wdt:P19/wdt:P131* wd:Q60") \%>\%
spq_add("?item wikibase:sitelinks ?linkcount") \%>\%
spq_arrange(desc(linkcount)) \%>\%
spq_head()

# Return 42 items
spq_init() \%>\%
spq_add("?item wdt:P31 wd:Q5",label="?item") \%>\%
spq_add("?item wdt:P19/wdt:P131* wd:Q60") \%>\%
spq_add("?item wikibase:sitelinks ?linkcount") \%>\%
spq_arrange(desc(linkcount)) \%>\%
spq_head(42)
}
