% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_init.R
\name{spq_init}
\alias{spq_init}
\title{Initialize a query object.}
\usage{
spq_init(
  request_control = spq_control_request(endpoint = "wikidata", user_agent =
    getOption("glitter.ua", "glitter R package (https://github.com/lvaudor/glitter)"),
    max_tries = getOption("glitter.max_tries", 3L), max_seconds =
    getOption("glitter.max_seconds", 120L), timeout = getOption("glitter.timeout",
    1000L), request_type = c("url", "body-form"))
)
}
\arguments{
\item{request_control}{An object as returned by \code{\link[=spq_control_request]{spq_control_request()}}}
}
\value{
A query object
}
\description{
Initialize a query object.
}
\section{Printing}{

SPARQL queries are shown using the cli package,
with a built-in theme.
You can change it by using the \code{cli.user_theme} option.
We use
\itemize{
\item \code{.emph} for keywords and functions,
\item \code{.field} for variables,
\item \code{.pkg} for prefixes,
\item \code{.val} for strings,
\item \code{.url} for prefix URLs.
}

You can also turn off the cli behavior by setting the environment variable
\code{"GLITTER.NOCLI"} to any non-empty string.
That's what we do in glitter snapshot tests.
}

