% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_label.R
\name{spq_label}
\alias{spq_label}
\title{Label variables}
\usage{
spq_label(.query, ..., .languages = getOption("glitter.lang", "en$"))
}
\arguments{
\item{.query}{a list with elements of the query}

\item{...}{variables by which to arrange
(or SPARQL strings escaped with \code{spq()}, or strings, see examples)}

\item{.languages}{Languages for which to query labels. If you write "en" you
can get labels for regional variants such as "en-GB". If you want results for
"en" only, write "en$".}
}
\value{
A query object
}
\description{
Label variables
}
\examples{
\dontrun{
spq_init() \%>\%
  spq_add("?mayor wdt:P31 ?species") \%>\%
# dog, cat or chicken
  spq_set(species = c('wd:Q144','wd:Q146', 'wd:Q780')) \%>\%
# who occupy the function
  spq_add("?mayor p:P39 ?node") \%>\%
# of mayor
  spq_add("?node ps:P39 wd:Q30185") \%>\%
# of some places
  spq_add("?node pq:P642 ?place") \%>\%
  spq_label(mayor, place, .languages = c("fr", "en", "de")) \%>\%
  spq_perform()
}
}
