% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_language.R
\name{spq_language}
\alias{spq_language}
\title{Get labels in a specified language (apply only to Wikidata)}
\usage{
spq_language(query = NULL, language = "en")
}
\arguments{
\item{query}{the query}

\item{language}{the language in which the labels will be provided (defaults to "en" for English). See complete list of Wikimedia language codes [here](https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all). You can also set language to "auto" and then the Wikidata SPARQL engine will try and detect your language automatically. Specifying several languages will return labels with languages following the priority specified (e.g. with language="fr,en", the label will be returned preferentially in French, or, if there is not French label for the item, in English).}
}
\description{
Get labels in a specified language (apply only to Wikidata)
}
\examples{
tib=spq_init() \%>\%
 spq_add("?film wdt:P31 wd:Q11424", label="?film") \%>\%
 spq_language("fr,en") \%>\%
 spq_head(10) \%>\%
 send()

tib=spq_init() \%>\%
 spq_add("?film wdt:P31 wd:Q11424", label="?film") \%>\%
 spq_language("auto,es,it,de") \%>\%
 spq_head(10) \%>\%
 send()
}
