% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_mutate.R
\name{spq_mutate}
\alias{spq_mutate}
\title{Create and modify variables in the results}
\usage{
spq_mutate(
  .query,
  ...,
  .label = NA,
  .within_box = c(NA, NA),
  .within_distance = c(NA, NA)
)
}
\arguments{
\item{.query}{a list with elements of the query}

\item{...}{variables by which to arrange
(or SPARQL strings escaped with \code{spq()}, or strings, see examples)}

\item{.label}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} See \code{\link[=spq_label]{spq_label()}}.}

\item{.within_box}{if provided, rectangular bounding box for the triple query.
Provided as list(southwest=c(long=...,lat=...),northeast=c(long=...,lat=...))}

\item{.within_distance}{if provided, circular bounding box for the triple query.
Provided as list(center=c(long=...,lat=...), radius=...),
with radius in kilometers.
The center can also be provided as a variable (for instance, "?location")
for the center coordinates to be retrieved directly from the query.}
}
\value{
A query object
}
\description{
Create and modify variables in the results
}
\section{Some examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{# common name of a plant species in different languages
# the triplet pattern "wd:Q331676 wdt:P1843 ?statement"
# creates the variable statement
# hence our writing it in reverse within the spq_mutate() function
spq_init() \%>\%
spq_mutate(statement = wdt::P1843(wd::Q331676)) \%>\%
spq_mutate(lang = lang(statement))
}\if{html}{\out{</div>}}
}

