% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_mutate.R
\name{spq_mutate}
\alias{spq_mutate}
\title{Create and modify variables in the results}
\usage{
spq_mutate(query, vars)
}
\arguments{
\item{query}{a list with elements of the query}

\item{name}{name of variable}

\item{description}{formula}
}
\description{
Create and modify variables in the results
}
\examples{
# common name of a plant species in different languages
spq_init() \%>\%
spq_add("wd:Q331676 wdt:P1843 ?statement") \%>\%
spq_mutate(c("?lang"="LANG(?statement)")) \%>\%
send()
}
