% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_prefix.R
\name{spq_prefix}
\alias{spq_prefix}
\title{Add prefixes to the query}
\usage{
spq_prefix(query = NULL, auto = TRUE, prefixes = NULL)
}
\arguments{
\item{query}{a list with elements of the query}

\item{prefixes}{a vector of prefixes}
}
\description{
Add prefixes to the query
}
\examples{
spq_init() \%>\%
spq_prefix(prefixes=c(dbo="http://dbpedia.org/ontology/"))
}
