% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_select.R
\name{spq_select}
\alias{spq_select}
\title{Select particular variables}
\usage{
spq_select(query = NULL, variables)
}
\arguments{
\item{query}{the query}

\item{variables}{the variables}
}
\description{
Select particular variables
}
\examples{
tib=spq_init() \%>\%
 spq_add("?film wdt:P31 wd:Q11424", label="?film") \%>\%
 spq_add("?film wdt:P840 ?loc", label="?loc") \%>\%
 spq_add("?film wdt:P495 ?origin", label="?origin") \%>\%
 spq_add("?film wdt:P577 ?date") \%>\%
 spq_mutate(c("?year"="year(?date)")) \%>\%
 spq_head(10) \%>\%
 spq_select("-?date") \%>\%
 send()
}
