% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_select.R
\name{spq_select}
\alias{spq_select}
\title{Select (and create) particular variables}
\usage{
spq_select(.query = NULL, ..., .spq_duplicate = NULL)
}
\arguments{
\item{.query}{a list with elements of the query}

\item{...}{variables by which to arrange
(or SPARQL strings escaped with \code{spq()}, or strings, see examples)}

\item{.spq_duplicate}{How to handle duplicates: keep them (\code{NULL}), eliminate (\code{distinct})
or reduce them (\code{reduced}, advanced usage).}
}
\value{
A query object
}
\description{
Select (and create) particular variables
}
\examples{

spq_init() |>
  spq_prefix(prefixes = c(dct = "http://purl.org/dc/terms/")) |>
  spq_add(spq('?lexemeId dct:language wd:Q1860')) |>
  spq_add(spq("?lexemeId wikibase:lemma ?lemma")) |>
  spq_filter(str_detect(lemma, '^pota.*')) |>
  spq_select(- lemma)

spq_init() |>
  spq_prefix(prefixes = c(dct = "http://purl.org/dc/terms/")) |>
  spq_add(spq('?lexemeId dct:language wd:Q1860')) |>
  spq_add(spq("?lexemeId wikibase:lemma ?lemma")) |>
  spq_filter(str_detect(lemma, '^pota.*')) |>
  spq_select(lemma)
}
