% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_select.R
\name{spq_select}
\alias{spq_select}
\title{Select (and create) particular variables}
\usage{
spq_select(.query = NULL, ..., .spq_duplicate = NULL)
}
\arguments{
\item{.query}{a list with elements of the query}

\item{...}{variables by which to arrange
(or SPARQL strings escaped with \code{spq()}, or strings, see examples)}

\item{.spq_duplicate}{How to handle duplicates: keep them (\code{NULL}), eliminate (\code{distinct})
or reduce them (\code{reduced}, advanced usage).}
}
\description{
Select (and create) particular variables
}
\examples{

query = spq_init()
spq_select(query, count = n (human), eyecolorLabel, haircolorLabel)
}
