% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_set.R
\name{spq_set}
\alias{spq_set}
\title{Set helper values for the query}
\usage{
spq_set(query, ...)
}
\arguments{
\item{query}{query}

\item{...}{Helper values and their definition.}
}
\value{

}
\description{
Set helper values for the query (helps with readability)
}
\examples{
# find the individuals of the species
spq_init() \%>\%
# dog, cat or chicken
spq_set(species = c('wd:144','wd:146', 'wd:780'), mayor = "Q30185") \%>\%
spq_add(spq("?mayor wdt:P31 ?species"))
}
