% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_set.R
\name{spq_set}
\alias{spq_set}
\title{Set helper values for the query}
\usage{
spq_set(.query, ...)
}
\arguments{
\item{.query}{query}

\item{...}{Helper values and their definition.}
}
\value{

}
\description{
Set helper values for the query (helps with readability)
}
\examples{
# find the individuals of the species
spq_init() \%>\%
# dog, cat or chicken
spq_set(species = c('wd:Q144','wd:Q146', 'wd:Q780'), mayorcode = "wd:Q30185") \%>\%
spq_filter(mayor == wdt::P31(species), .label = "?species") \%>\%
spq_add("?mayor p:P39 ?node", .label = "?mayor") \%>\%
# of mayor
spq_add("?node ps:P39 ?mayorcode") \%>\%
# of some places
spq_add("?node pq:P642 ?place", .label = "?place") \%>\%
spq_select(-species, -place, -node, -mayor, -mayorcode) \%>\%
spq_perform()
}
