% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_summarise.R
\name{spq_summarise}
\alias{spq_summarise}
\title{Summarise each group of results to fewer results}
\usage{
spq_summarise(query, vars)
}
\arguments{
\item{query}{a list with elements of the query}

\item{vars}{a vector with the names of the variables used for grouping}
}
\description{
Summarise each group of results to fewer results
}
\examples{
result=spq_init() \%>\%
add_triplets("?item wdt:P361 wd:Q297853") \%>\%
add_triplets("?item wdt:P1082 ?folkm_ngd") \%>\%
add_triplets("?area wdt:P31 wd:Q1907114",label="?area") \%>\%
add_triplets("?area wdt:P527 ?item") \%>\%
spq_group_by(c("?area","?areaLabel")) \%>\%
spq_summarise(c("?total_folkm"="sum(?folkm_ngd)")) \%>\%
send()
}
