% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_summarise.R
\name{spq_summarise}
\alias{spq_summarise}
\alias{spq_summarize}
\title{Summarise each group of results to fewer results}
\usage{
spq_summarise(query, ...)

spq_summarize(query, ...)
}
\arguments{
\item{query}{a list with elements of the query}

\item{...}{variables by which to arrange
(or SPARQL strings escaped with \code{spq()}, or strings, see examples)}
}
\description{
Summarise each group of results to fewer results
}
\examples{
result=spq_init() \%>\%
spq_add("?item wdt:P361 wd:Q297853") \%>\%
spq_add("?item wdt:P1082 ?folkm_ngd") \%>\%
spq_add("?area wdt:P31 wd:Q1907114",label="?area") \%>\%
spq_add("?area wdt:P527 ?item") \%>\%
spq_group_by(c("?area","?areaLabel"))  \%>\%
spq_summarise(total_folkm = sum(folkm_ngd))
}
