% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{status}
\alias{status}
\title{Get the Git status of any local repo using the path}
\usage{
status(path = getwd(), verbose = TRUE, header = "Status Response")
}
\arguments{
\item{path}{full path to local repository to be pushed}
}
\value{
If the git message is of a length greater than 0, it is returned as a character vector and also printed in the console
}
\description{
Get the Git status of any local repo using the path
}
