% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{transform_wikidata_coords}
\alias{transform_wikidata_coords}
\title{Get coords at lat-lng from Wikidata format}
\usage{
transform_wikidata_coords(data, coord_column, prefix = "")
}
\arguments{
\item{data}{a tibble}

\item{coord_column}{the name of the column with coordinates formatted as 'Point(..... ....)'}

\item{prefix}{the prefix to be appended to "lat-lng" columns (defaults to "" -no prefix-)}
}
\description{
Get coords at lat-lng from Wikidata format
}
\examples{
query='SELECT ?item ?itemLabel ?coords ?date
{
  ?item wdt:P361 wd:Q1552;
        wdt:P625 ?coords.
  OPTIONAL{?item wdt:P1619 ?date.}
  SERVICE wikibase:label { bd:serviceParam wikibase:language "en" . }
} ORDER BY ?itemLabel
'
tib=send_query(query)
transform_wikidata_coords(tib, "coords")
}
