% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untracked_files.R
\name{untracked_files}
\alias{untracked_files}
\title{Get a list of untracked files in a repo}
\usage{
untracked_files(path_to_local_repo)
}
\arguments{
\item{path_to_local_repo}{path to local repo}
}
\value{
The complete git status message and untracked files are printed in the console, and a vector of the untracked filenames is invisibly returned.
}
\description{
This function takes the git status message and isolates the files that are new/untracked according to that message.
}
