% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untracked_files.R
\name{untracked_files}
\alias{untracked_files}
\title{List all unstaged files by File Type}
\usage{
untracked_files(path_to_local_repo = NULL)
}
\description{
List all unstaged files by File Type
}
\keyword{internal}
