% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-files.R
\name{untracked_files}
\alias{untracked_files}
\title{(Deprecated) List all unstaged files by File Type}
\usage{
untracked_files(path = getwd())
}
\description{
(Deprecated) List all unstaged files by File Type
}
