% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_files_to_commit.R
\name{wd_files_to_commit}
\alias{wd_files_to_commit}
\title{Get a list of both untracked and modified files in the wd only}
\usage{
wd_files_to_commit()
}
\description{
Get a list of both untracked and modified files in the wd only
}
